amv_dimnames {plyr}	R Documentation

Dimension names.

Description

Consistent dimnames for vectors, matrices and arrays.

Usage

amv_dimnames(x)

Arguments

x
array, matrix or vector

Details

Unlike dimnames no part of the output will ever be null. If a component of dimnames is omitted, amv_dimnames will return an integer sequence of the appropriate length.

[Package plyr version 1.8.9 Index]