as.data.frame.function {plyr}	R Documentation

Make a function return a data frame.

Description

Create a new function that returns the existing function wrapped in a data.frame

Usage

  ## S3 method for class 'function'
 as.data.frame(x, row.names, optional,
    ...)

Arguments

x
function to make return a data frame

row.names
necessary to match the generic, but not used

optional
necessary to match the generic, but not used

...
necessary to match the generic, but not used

Details

This is useful when calling *dply functions with a function that returns a vector, and you want the output in rows, rather than columns

[Package plyr version 1.8 Index]