as.quoted {plyr}	R Documentation

Convert input to quoted variables.

Description

Convert characters, formulas and calls to quoted .variables

Usage

as.quoted(x, env = parent.frame())

Arguments

x
input to quote

env
environment in which unbound symbols in expression should be evaluated. Defaults to the environment in which as.quoted was executed.

Details

This method is called by default on all plyr functions that take a .variables argument, so that equivalent forms can be used anywhere.

Currently conversions exist for character vectors, formulas and call objects.

Value

a list of quoted variables

See Also

.

Examples

as.quoted(c("a", "b", "log(d)"))
as.quoted(a ~ b + log(d))

[Package plyr version 1.8.7 Index]