Split data frame, apply function, and return results in an array.

For each subset of data frame, apply function then combine results into an array. daply with a function that operates column-wise is similar to aggregate. To apply a function for each row, use aaply with .margins set to 1.

daply(.data, .variables, .fun = NULL, ..., .progress = "none",
  .inform = FALSE, .drop_i = TRUE, .drop_o = TRUE, .parallel = FALSE,
  .paropts = NULL)

Arguments

.data
data frame to be processed

.variables
variables to split data frame by, as quoted variables, a formula or character vector

.fun
function to apply to each piece

...
other arguments passed on to .fun

.progress
name of the progress bar to use, see create_progress_bar

.inform
produce informative error messages? This is turned off by default because it substantially slows processing speed, but is very useful for debugging

.drop_i
should combinations of variables that do not appear in the input data be preserved (FALSE) or dropped (TRUE, default)

.drop_o
should extra dimensions of length 1 in the output be dropped, simplifying the output. Defaults to TRUE

.parallel
if TRUE, apply function in parallel, using parallel backend provided by foreach

.paropts
a list of additional options passed into the foreach function when parallel computation is enabled. This is important if (for example) your code relies on external data or packages: use the .export and .packages arguments to supply them so that all cluster nodes have the correct environment set up for computing.

Value

if results are atomic with same type and dimensionality, a vector, matrix or array; otherwise, a list-array (a list with dimensions)

Input

This function splits data frames by variables.

Output

If there are no results, then this function will return a vector of length 0 (vector()).

References

Hadley Wickham (2011). The Split-Apply-Combine Strategy for Data Analysis. Journal of Statistical Software, 40(1), 1-29. http://www.jstatsoft.org/v40/i01/.

See also

Examples

daply(baseball, .(year), nrow)

#> 1871 1872 1873 1874 1875 1876 1877 1878 1879 1880 1881 1882 1883 1884 1885 1886 
#>    7   13   13   15   17   15   17   17   25   28   36   35   38   53   50   51 
#> 1887 1888 1889 1890 1891 1892 1893 1894 1895 1896 1897 1898 1899 1900 1901 1902 
#>   56   60   57   72   77   75   68   67   70   74   75   79   77   69   83   87 
#> 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915 1916 1917 1918 
#>   79   81   82   81   82   81   82   75   86   99  101  108  133  122  112  110 
#> 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932 1933 1934 
#>  124  117  117  115  133  132  137  136  134  133  132  133  125  132  133  133 
#> 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947 1948 1949 1950 
#>  127  134  121  118  130  122  132  120  101   90   89  122  124  112  120  120 
#> 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 
#>  126  143  141  139  162  173  173  175  184  199  198  192  200  219  218  230 
#> 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 
#>  227  214  253  268  274  282  295  300  308  315  323  309  319  298  307  316 
#> 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992 1993 1994 1995 1996 1997 1998 
#>  318  316  315  329  330  326  339  333  333  351  355  333  363  343  335  324 
#> 1999 2000 2001 2002 2003 2004 2005 2006 2007 
#>  283  285  261  213  216  187  160  141   92 

# Several different ways of summarising by variables that should not be
# included in the summary

daply(baseball[, c(2, 6:9)], .(year), colwise(mean))

#>       
#> year   g        ab       r        h       
#>   1871 28       135.8571 33.57143 42.14286
#>   1872 29.46154 140.8462 32.15385 42.92308
#>   1873 46.30769 217.6154 48.46154 68.53846
#>   1874 49       226.7333 44       64.86667
#>   1875 57.82353 256.9412 47.17647 73.29412
#>   1876 58.66667 258.4    43.26667 72.4    
#>   1877 49.58824 212.2941 36.58824 64.23529
#>   1878 56.17647 232.8235 36.82353 66.82353
#>   1879 68.92    293.6    48.32    82.32   
#>   1880 67.35714 272.5357 41.07143 72      
#>   1881 61.52778 247.75   38.44444 68.41667
#>   1882 73.65714 305.0286 52.2     85.42857
#>   1883 90.71053 379.7895 70.94737 108.7632
#>   1884 80.28302 327.2075 63.32075 92.15094
#>   1885 77.74    312.34   55.4     83.9    
#>   1886 87.96078 343.8235 65.62745 95.96078
#>   1887 83.57143 332.7143 65.80357 95.85714
#>   1888 79.5     309.6833 47.25    80.36667
#>   1889 93.03509 364.8596 68.38596 102.7193
#>   1890 89.70833 353.3611 68.15278 99.38889
#>   1891 79.96104 310.2597 54.25974 83.96104
#>   1892 88.81333 343.2    54.41333 88.78667
#>   1893 83.98529 336.8382 65.48529 100.1029
#>   1894 84.73134 338.0896 76.07463 110.194 
#>   1895 83.14286 327.2714 64.67143 101.8857
#>   1896 78.27027 300.473  56.05405 93.04054
#>   1897 78.13333 299.7067 53.58667 93.54667
#>   1898 86.81013 324.2911 50.26582 92.68354
#>   1899 84.62338 312.4805 51.93506 93.06494
#>   1900 77.85507 287.6087 45.27536 83.26087
#>   1901 77.04819 290.6024 43.45783 84.06024
#>   1902 78.88506 294.4483 41.32184 83.7931 
#>   1903 80.07595 293.2152 41.96203 83.74684
#>   1904 85.02469 303.7901 36.74074 79.50617
#>   1905 87.67073 312.1585 40.0122  83      
#>   1906 82.60494 289.5309 35.66667 76.75309
#>   1907 76.58537 262.6707 31.96341 68.9878 
#>   1908 80.75309 269.963  32.90123 68.83951
#>   1909 74.65854 250.7927 31.86585 65.45122
#>   1910 85.02667 293.2533 41       81.08   
#>   1911 79.76744 272.0465 42.04651 77.12791
#>   1912 67.14141 225.5253 34.06061 64.46465
#>   1913 72.72277 234.9307 31.50495 63.86139
#>   1914 83.39815 269.5185 34.01852 71.7963 
#>   1915 71.4812  232.6391 29.11278 60.71429
#>   1916 75.44262 243.6639 28.87705 64.60656
#>   1917 81.75893 272.1518 32.49107 71.97321
#>   1918 66.66364 229.0091 26.56364 60.50909
#>   1919 69.28226 232.9919 29.29839 64.70161
#>   1920 87.10256 299.094  40.70085 86.32479
#>   1921 83.42735 288.5983 43.30769 87.81197
#>   1922 86.97391 297.0783 43.77391 89.46957
#>   1923 78.2406  262      38.45865 78      
#>   1924 81.71212 272.6818 39.67424 82.06818
#>   1925 68.36496 222.2482 34.16788 67.91971
#>   1926 71.67647 229.1324 32.58088 67.25735
#>   1927 69.91791 224.7761 32.79104 66.48507
#>   1928 71.09023 224.7895 32.44361 65.55639
#>   1929 71.15152 231.2727 36.31061 69.4697 
#>   1930 68.07519 221.9398 36.97744 68.28571
#>   1931 72.704   237.424  34.28    67.984  
#>   1932 69.86364 226.5303 33.85606 64.41667
#>   1933 71.17293 232.3459 31.49624 65.19549
#>   1934 72.26316 234.7895 33.66917 67.66165
#>   1935 72.38583 241.4488 34.89764 68.99213
#>   1936 72.30597 243.5672 36.77612 71.47015
#>   1937 71.81818 235.7521 34.78512 67.84298
#>   1938 67.72881 223.4576 32.68644 62.94915
#>   1939 63       202.9846 29.54615 57.14615
#>   1940 66.90164 215.8033 30.16393 58.64754
#>   1941 61.05303 194.7803 26.11364 52.7803 
#>   1942 63.49167 203      25.48333 52.90833
#>   1943 66.76238 214.4653 25.76238 57.39604
#>   1944 65.58889 214.9556 28.14444 60.04444
#>   1945 60.60674 201.5281 26.1236  55.40449
#>   1946 58.57377 185.8197 23.27869 50.79508
#>   1947 64.96774 204.5726 28.21774 55.78226
#>   1948 77.17857 244.6696 34.71429 67.35714
#>   1949 78.95    258.8583 37.63333 71.73333
#>   1950 80.70833 269.0833 40.175   75.45833
#>   1951 79.53968 261.3571 37.19841 71.39683
#>   1952 70.55245 227.021  29.65035 59.76224
#>   1953 74       235.0071 33.70922 65.02128
#>   1954 76.33813 239.6187 33.84892 66.41727
#>   1955 70.05556 213.1975 31.18519 58.48765
#>   1956 71.3815  217.8671 31.09827 59.49133
#>   1957 68.50289 206.052  27.86705 55.78035
#>   1958 68.61143 196.7314 26.42857 53.68   
#>   1959 65.72826 189.538  26.41304 50.94022
#>   1960 63.32663 181.1206 24.34171 48.26633
#>   1961 67.18182 194.6869 27.31818 52.94444
#>   1962 70.97917 211.5312 29.13542 56.73958
#>   1963 72.045   213.48   26.485   54.99   
#>   1964 69.3242  204.0274 26.24201 53.88584
#>   1965 71.97248 208.7661 26.52752 54.04587
#>   1966 70.18696 205.3783 25.87826 54.05217
#>   1967 71.837   212.0749 26.09251 54.5815 
#>   1968 75.97664 225.6402 25.19626 56.33645
#>   1969 73.26877 220.6759 29.23715 57.89328
#>   1970 72.12313 216.306  30.00373 57.89179
#>   1971 74.0073  225.4088 27.91241 58.76642
#>   1972 72.96809 218.234  26.6383  56.10993
#>   1973 78.7322  236.2034 30.76949 63.02373
#>   1974 80.69    235.9133 30.86333 63.32   
#>   1975 80.75325 238.2695 31.46104 64.00325
#>   1976 77.92381 230.7175 29.06032 60.93968
#>   1977 76.56037 227.4396 31.82043 61.87616
#>   1978 82.16181 240.3528 31.18447 63.94822
#>   1979 78.89655 232.8652 32.52665 63.89655
#>   1980 82.25168 241.7617 32.28188 66.30201
#>   1981 54.47883 157.3518 19.85668 41.61238
#>   1982 81.53481 233.8481 31.5538  63.2943 
#>   1983 78.37107 225.6226 30.22013 61.05031
#>   1984 76.76582 218.7658 28.3038  58.96203
#>   1985 76.39683 219.7079 29.6381  58.44444
#>   1986 74.25532 210.848  28.08207 55.80243
#>   1987 71.03333 200.6212 29.22424 54.38182
#>   1988 69.95399 197.8865 25.76074 51.80368
#>   1989 68.76401 191.056  23.9587  49.61062
#>   1990 73.45045 208.5736 27.22523 55.57658
#>   1991 75.85886 214.1802 28.64565 56.61562
#>   1992 71.94302 203.0285 26       53.70085
#>   1993 73.42817 209.293  29.8169  57.84789
#>   1994 57.35736 167.6757 25.69369 46.84084
#>   1995 64.25344 186.5317 28.03581 51.90358
#>   1996 74.62682 217.3673 33.29155 60.97959
#>   1997 72.04478 212.2925 31.26269 58.58209
#>   1998 72.49074 216.1574 32.53395 60.10494
#>   1999 75.07774 220.1272 35.25795 62.05654
#>   2000 65.58596 201.9439 31.57193 56.16491
#>   2001 66.45977 195.7011 29.35249 53.0728 
#>   2002 69.51643 197.7653 28.00469 52.97653
#>   2003 59.51389 167.4815 24.4213  44.71759
#>   2004 61.65241 171.754  24.90374 46.31016
#>   2005 64.8625  158.1813 20.6875  41.4125 
#>   2006 55.28369 130.3262 17.36879 34.04255
#>   2007 51.66304 136.4674 18.5     35.83696

daply(baseball[, 6:9], .(baseball$year), colwise(mean))

#>              
#> baseball$year g        ab       r        h       
#>          1871 28       135.8571 33.57143 42.14286
#>          1872 29.46154 140.8462 32.15385 42.92308
#>          1873 46.30769 217.6154 48.46154 68.53846
#>          1874 49       226.7333 44       64.86667
#>          1875 57.82353 256.9412 47.17647 73.29412
#>          1876 58.66667 258.4    43.26667 72.4    
#>          1877 49.58824 212.2941 36.58824 64.23529
#>          1878 56.17647 232.8235 36.82353 66.82353
#>          1879 68.92    293.6    48.32    82.32   
#>          1880 67.35714 272.5357 41.07143 72      
#>          1881 61.52778 247.75   38.44444 68.41667
#>          1882 73.65714 305.0286 52.2     85.42857
#>          1883 90.71053 379.7895 70.94737 108.7632
#>          1884 80.28302 327.2075 63.32075 92.15094
#>          1885 77.74    312.34   55.4     83.9    
#>          1886 87.96078 343.8235 65.62745 95.96078
#>          1887 83.57143 332.7143 65.80357 95.85714
#>          1888 79.5     309.6833 47.25    80.36667
#>          1889 93.03509 364.8596 68.38596 102.7193
#>          1890 89.70833 353.3611 68.15278 99.38889
#>          1891 79.96104 310.2597 54.25974 83.96104
#>          1892 88.81333 343.2    54.41333 88.78667
#>          1893 83.98529 336.8382 65.48529 100.1029
#>          1894 84.73134 338.0896 76.07463 110.194 
#>          1895 83.14286 327.2714 64.67143 101.8857
#>          1896 78.27027 300.473  56.05405 93.04054
#>          1897 78.13333 299.7067 53.58667 93.54667
#>          1898 86.81013 324.2911 50.26582 92.68354
#>          1899 84.62338 312.4805 51.93506 93.06494
#>          1900 77.85507 287.6087 45.27536 83.26087
#>          1901 77.04819 290.6024 43.45783 84.06024
#>          1902 78.88506 294.4483 41.32184 83.7931 
#>          1903 80.07595 293.2152 41.96203 83.74684
#>          1904 85.02469 303.7901 36.74074 79.50617
#>          1905 87.67073 312.1585 40.0122  83      
#>          1906 82.60494 289.5309 35.66667 76.75309
#>          1907 76.58537 262.6707 31.96341 68.9878 
#>          1908 80.75309 269.963  32.90123 68.83951
#>          1909 74.65854 250.7927 31.86585 65.45122
#>          1910 85.02667 293.2533 41       81.08   
#>          1911 79.76744 272.0465 42.04651 77.12791
#>          1912 67.14141 225.5253 34.06061 64.46465
#>          1913 72.72277 234.9307 31.50495 63.86139
#>          1914 83.39815 269.5185 34.01852 71.7963 
#>          1915 71.4812  232.6391 29.11278 60.71429
#>          1916 75.44262 243.6639 28.87705 64.60656
#>          1917 81.75893 272.1518 32.49107 71.97321
#>          1918 66.66364 229.0091 26.56364 60.50909
#>          1919 69.28226 232.9919 29.29839 64.70161
#>          1920 87.10256 299.094  40.70085 86.32479
#>          1921 83.42735 288.5983 43.30769 87.81197
#>          1922 86.97391 297.0783 43.77391 89.46957
#>          1923 78.2406  262      38.45865 78      
#>          1924 81.71212 272.6818 39.67424 82.06818
#>          1925 68.36496 222.2482 34.16788 67.91971
#>          1926 71.67647 229.1324 32.58088 67.25735
#>          1927 69.91791 224.7761 32.79104 66.48507
#>          1928 71.09023 224.7895 32.44361 65.55639
#>          1929 71.15152 231.2727 36.31061 69.4697 
#>          1930 68.07519 221.9398 36.97744 68.28571
#>          1931 72.704   237.424  34.28    67.984  
#>          1932 69.86364 226.5303 33.85606 64.41667
#>          1933 71.17293 232.3459 31.49624 65.19549
#>          1934 72.26316 234.7895 33.66917 67.66165
#>          1935 72.38583 241.4488 34.89764 68.99213
#>          1936 72.30597 243.5672 36.77612 71.47015
#>          1937 71.81818 235.7521 34.78512 67.84298
#>          1938 67.72881 223.4576 32.68644 62.94915
#>          1939 63       202.9846 29.54615 57.14615
#>          1940 66.90164 215.8033 30.16393 58.64754
#>          1941 61.05303 194.7803 26.11364 52.7803 
#>          1942 63.49167 203      25.48333 52.90833
#>          1943 66.76238 214.4653 25.76238 57.39604
#>          1944 65.58889 214.9556 28.14444 60.04444
#>          1945 60.60674 201.5281 26.1236  55.40449
#>          1946 58.57377 185.8197 23.27869 50.79508
#>          1947 64.96774 204.5726 28.21774 55.78226
#>          1948 77.17857 244.6696 34.71429 67.35714
#>          1949 78.95    258.8583 37.63333 71.73333
#>          1950 80.70833 269.0833 40.175   75.45833
#>          1951 79.53968 261.3571 37.19841 71.39683
#>          1952 70.55245 227.021  29.65035 59.76224
#>          1953 74       235.0071 33.70922 65.02128
#>          1954 76.33813 239.6187 33.84892 66.41727
#>          1955 70.05556 213.1975 31.18519 58.48765
#>          1956 71.3815  217.8671 31.09827 59.49133
#>          1957 68.50289 206.052  27.86705 55.78035
#>          1958 68.61143 196.7314 26.42857 53.68   
#>          1959 65.72826 189.538  26.41304 50.94022
#>          1960 63.32663 181.1206 24.34171 48.26633
#>          1961 67.18182 194.6869 27.31818 52.94444
#>          1962 70.97917 211.5312 29.13542 56.73958
#>          1963 72.045   213.48   26.485   54.99   
#>          1964 69.3242  204.0274 26.24201 53.88584
#>          1965 71.97248 208.7661 26.52752 54.04587
#>          1966 70.18696 205.3783 25.87826 54.05217
#>          1967 71.837   212.0749 26.09251 54.5815 
#>          1968 75.97664 225.6402 25.19626 56.33645
#>          1969 73.26877 220.6759 29.23715 57.89328
#>          1970 72.12313 216.306  30.00373 57.89179
#>          1971 74.0073  225.4088 27.91241 58.76642
#>          1972 72.96809 218.234  26.6383  56.10993
#>          1973 78.7322  236.2034 30.76949 63.02373
#>          1974 80.69    235.9133 30.86333 63.32   
#>          1975 80.75325 238.2695 31.46104 64.00325
#>          1976 77.92381 230.7175 29.06032 60.93968
#>          1977 76.56037 227.4396 31.82043 61.87616
#>          1978 82.16181 240.3528 31.18447 63.94822
#>          1979 78.89655 232.8652 32.52665 63.89655
#>          1980 82.25168 241.7617 32.28188 66.30201
#>          1981 54.47883 157.3518 19.85668 41.61238
#>          1982 81.53481 233.8481 31.5538  63.2943 
#>          1983 78.37107 225.6226 30.22013 61.05031
#>          1984 76.76582 218.7658 28.3038  58.96203
#>          1985 76.39683 219.7079 29.6381  58.44444
#>          1986 74.25532 210.848  28.08207 55.80243
#>          1987 71.03333 200.6212 29.22424 54.38182
#>          1988 69.95399 197.8865 25.76074 51.80368
#>          1989 68.76401 191.056  23.9587  49.61062
#>          1990 73.45045 208.5736 27.22523 55.57658
#>          1991 75.85886 214.1802 28.64565 56.61562
#>          1992 71.94302 203.0285 26       53.70085
#>          1993 73.42817 209.293  29.8169  57.84789
#>          1994 57.35736 167.6757 25.69369 46.84084
#>          1995 64.25344 186.5317 28.03581 51.90358
#>          1996 74.62682 217.3673 33.29155 60.97959
#>          1997 72.04478 212.2925 31.26269 58.58209
#>          1998 72.49074 216.1574 32.53395 60.10494
#>          1999 75.07774 220.1272 35.25795 62.05654
#>          2000 65.58596 201.9439 31.57193 56.16491
#>          2001 66.45977 195.7011 29.35249 53.0728 
#>          2002 69.51643 197.7653 28.00469 52.97653
#>          2003 59.51389 167.4815 24.4213  44.71759
#>          2004 61.65241 171.754  24.90374 46.31016
#>          2005 64.8625  158.1813 20.6875  41.4125 
#>          2006 55.28369 130.3262 17.36879 34.04255
#>          2007 51.66304 136.4674 18.5     35.83696

daply(baseball, .(year), function(df) colwise(mean)(df[, 6:9]))

#>       
#> year   g        ab       r        h       
#>   1871 28       135.8571 33.57143 42.14286
#>   1872 29.46154 140.8462 32.15385 42.92308
#>   1873 46.30769 217.6154 48.46154 68.53846
#>   1874 49       226.7333 44       64.86667
#>   1875 57.82353 256.9412 47.17647 73.29412
#>   1876 58.66667 258.4    43.26667 72.4    
#>   1877 49.58824 212.2941 36.58824 64.23529
#>   1878 56.17647 232.8235 36.82353 66.82353
#>   1879 68.92    293.6    48.32    82.32   
#>   1880 67.35714 272.5357 41.07143 72      
#>   1881 61.52778 247.75   38.44444 68.41667
#>   1882 73.65714 305.0286 52.2     85.42857
#>   1883 90.71053 379.7895 70.94737 108.7632
#>   1884 80.28302 327.2075 63.32075 92.15094
#>   1885 77.74    312.34   55.4     83.9    
#>   1886 87.96078 343.8235 65.62745 95.96078
#>   1887 83.57143 332.7143 65.80357 95.85714
#>   1888 79.5     309.6833 47.25    80.36667
#>   1889 93.03509 364.8596 68.38596 102.7193
#>   1890 89.70833 353.3611 68.15278 99.38889
#>   1891 79.96104 310.2597 54.25974 83.96104
#>   1892 88.81333 343.2    54.41333 88.78667
#>   1893 83.98529 336.8382 65.48529 100.1029
#>   1894 84.73134 338.0896 76.07463 110.194 
#>   1895 83.14286 327.2714 64.67143 101.8857
#>   1896 78.27027 300.473  56.05405 93.04054
#>   1897 78.13333 299.7067 53.58667 93.54667
#>   1898 86.81013 324.2911 50.26582 92.68354
#>   1899 84.62338 312.4805 51.93506 93.06494
#>   1900 77.85507 287.6087 45.27536 83.26087
#>   1901 77.04819 290.6024 43.45783 84.06024
#>   1902 78.88506 294.4483 41.32184 83.7931 
#>   1903 80.07595 293.2152 41.96203 83.74684
#>   1904 85.02469 303.7901 36.74074 79.50617
#>   1905 87.67073 312.1585 40.0122  83      
#>   1906 82.60494 289.5309 35.66667 76.75309
#>   1907 76.58537 262.6707 31.96341 68.9878 
#>   1908 80.75309 269.963  32.90123 68.83951
#>   1909 74.65854 250.7927 31.86585 65.45122
#>   1910 85.02667 293.2533 41       81.08   
#>   1911 79.76744 272.0465 42.04651 77.12791
#>   1912 67.14141 225.5253 34.06061 64.46465
#>   1913 72.72277 234.9307 31.50495 63.86139
#>   1914 83.39815 269.5185 34.01852 71.7963 
#>   1915 71.4812  232.6391 29.11278 60.71429
#>   1916 75.44262 243.6639 28.87705 64.60656
#>   1917 81.75893 272.1518 32.49107 71.97321
#>   1918 66.66364 229.0091 26.56364 60.50909
#>   1919 69.28226 232.9919 29.29839 64.70161
#>   1920 87.10256 299.094  40.70085 86.32479
#>   1921 83.42735 288.5983 43.30769 87.81197
#>   1922 86.97391 297.0783 43.77391 89.46957
#>   1923 78.2406  262      38.45865 78      
#>   1924 81.71212 272.6818 39.67424 82.06818
#>   1925 68.36496 222.2482 34.16788 67.91971
#>   1926 71.67647 229.1324 32.58088 67.25735
#>   1927 69.91791 224.7761 32.79104 66.48507
#>   1928 71.09023 224.7895 32.44361 65.55639
#>   1929 71.15152 231.2727 36.31061 69.4697 
#>   1930 68.07519 221.9398 36.97744 68.28571
#>   1931 72.704   237.424  34.28    67.984  
#>   1932 69.86364 226.5303 33.85606 64.41667
#>   1933 71.17293 232.3459 31.49624 65.19549
#>   1934 72.26316 234.7895 33.66917 67.66165
#>   1935 72.38583 241.4488 34.89764 68.99213
#>   1936 72.30597 243.5672 36.77612 71.47015
#>   1937 71.81818 235.7521 34.78512 67.84298
#>   1938 67.72881 223.4576 32.68644 62.94915
#>   1939 63       202.9846 29.54615 57.14615
#>   1940 66.90164 215.8033 30.16393 58.64754
#>   1941 61.05303 194.7803 26.11364 52.7803 
#>   1942 63.49167 203      25.48333 52.90833
#>   1943 66.76238 214.4653 25.76238 57.39604
#>   1944 65.58889 214.9556 28.14444 60.04444
#>   1945 60.60674 201.5281 26.1236  55.40449
#>   1946 58.57377 185.8197 23.27869 50.79508
#>   1947 64.96774 204.5726 28.21774 55.78226
#>   1948 77.17857 244.6696 34.71429 67.35714
#>   1949 78.95    258.8583 37.63333 71.73333
#>   1950 80.70833 269.0833 40.175   75.45833
#>   1951 79.53968 261.3571 37.19841 71.39683
#>   1952 70.55245 227.021  29.65035 59.76224
#>   1953 74       235.0071 33.70922 65.02128
#>   1954 76.33813 239.6187 33.84892 66.41727
#>   1955 70.05556 213.1975 31.18519 58.48765
#>   1956 71.3815  217.8671 31.09827 59.49133
#>   1957 68.50289 206.052  27.86705 55.78035
#>   1958 68.61143 196.7314 26.42857 53.68   
#>   1959 65.72826 189.538  26.41304 50.94022
#>   1960 63.32663 181.1206 24.34171 48.26633
#>   1961 67.18182 194.6869 27.31818 52.94444
#>   1962 70.97917 211.5312 29.13542 56.73958
#>   1963 72.045   213.48   26.485   54.99   
#>   1964 69.3242  204.0274 26.24201 53.88584
#>   1965 71.97248 208.7661 26.52752 54.04587
#>   1966 70.18696 205.3783 25.87826 54.05217
#>   1967 71.837   212.0749 26.09251 54.5815 
#>   1968 75.97664 225.6402 25.19626 56.33645
#>   1969 73.26877 220.6759 29.23715 57.89328
#>   1970 72.12313 216.306  30.00373 57.89179
#>   1971 74.0073  225.4088 27.91241 58.76642
#>   1972 72.96809 218.234  26.6383  56.10993
#>   1973 78.7322  236.2034 30.76949 63.02373
#>   1974 80.69    235.9133 30.86333 63.32   
#>   1975 80.75325 238.2695 31.46104 64.00325
#>   1976 77.92381 230.7175 29.06032 60.93968
#>   1977 76.56037 227.4396 31.82043 61.87616
#>   1978 82.16181 240.3528 31.18447 63.94822
#>   1979 78.89655 232.8652 32.52665 63.89655
#>   1980 82.25168 241.7617 32.28188 66.30201
#>   1981 54.47883 157.3518 19.85668 41.61238
#>   1982 81.53481 233.8481 31.5538  63.2943 
#>   1983 78.37107 225.6226 30.22013 61.05031
#>   1984 76.76582 218.7658 28.3038  58.96203
#>   1985 76.39683 219.7079 29.6381  58.44444
#>   1986 74.25532 210.848  28.08207 55.80243
#>   1987 71.03333 200.6212 29.22424 54.38182
#>   1988 69.95399 197.8865 25.76074 51.80368
#>   1989 68.76401 191.056  23.9587  49.61062
#>   1990 73.45045 208.5736 27.22523 55.57658
#>   1991 75.85886 214.1802 28.64565 56.61562
#>   1992 71.94302 203.0285 26       53.70085
#>   1993 73.42817 209.293  29.8169  57.84789
#>   1994 57.35736 167.6757 25.69369 46.84084
#>   1995 64.25344 186.5317 28.03581 51.90358
#>   1996 74.62682 217.3673 33.29155 60.97959
#>   1997 72.04478 212.2925 31.26269 58.58209
#>   1998 72.49074 216.1574 32.53395 60.10494
#>   1999 75.07774 220.1272 35.25795 62.05654
#>   2000 65.58596 201.9439 31.57193 56.16491
#>   2001 66.45977 195.7011 29.35249 53.0728 
#>   2002 69.51643 197.7653 28.00469 52.97653
#>   2003 59.51389 167.4815 24.4213  44.71759
#>   2004 61.65241 171.754  24.90374 46.31016
#>   2005 64.8625  158.1813 20.6875  41.4125 
#>   2006 55.28369 130.3262 17.36879 34.04255
#>   2007 51.66304 136.4674 18.5     35.83696