id {plyr}	R Documentation

Compute a unique numeric id for each unique row in a data frame.

Description

Properties:

order(id) is equivalent to do.call(order, df)

rows containing the same data have the same value

if drop = FALSE then room for all possibilites

Usage

id(.variables, drop = FALSE)

Arguments

.variables
list of variables

drop
drop unusued factor levels?

Value

a numeric vector with attribute n, giving total number of possibilities

See Also

id_var

[Package plyr version 1.8.7 Index]