indexed_df {plyr}	R Documentation

An indexed data frame.

Description

Create a indexed list, a space efficient way of indexing into a large data frame

Usage

indexed_df(data, index, vars)

Arguments

index
list of indices

vars
a character vector giving the variables used for subsetting

env
environment containing data frame

[Package plyr version 1.8.7 Index]