join.keys {plyr}	R Documentation

Join keys. Given two data frames, create a unique key for each row.

Description

Join keys. Given two data frames, create a unique key for each row.

Usage

  join.keys(x, y, by)

Arguments

x
data frame

y
data frame

by
character vector of variable names to join by

[Package plyr version 1.8 Index]