list_to_array {plyr}	R Documentation

List to array.

Description

Reduce/simplify a list of homogenous objects to an array

Usage

list_to_array(res, labels = NULL, .drop = FALSE)

Arguments

res
list of input data

labels
a data frame of labels, one row for each element of res

.drop
should extra dimensions be dropped (TRUE) or preserved (FALSE)

See Also

Other list simplification functions: list_to_dataframe(), list_to_vector()

[Package plyr version 1.8.9 Index]