name_rows {plyr}	R Documentation

Toggle row names between explicit and implicit.

Description

Plyr functions ignore row names, so this function provides a way to preserve them by converting them to an explicit column in the data frame. After the plyr operation, you can then apply name_rows again to convert back from the explicit column to the implicit rownames.

Usage

name_rows(df)

Arguments

df
a data.frame, with either rownames, or a column called .rownames.

Examples

name_rows(mtcars)
name_rows(name_rows(mtcars))

df <- data.frame(a = sample(10))
arrange(df, a)
arrange(name_rows(df), a)
name_rows(arrange(name_rows(df), a))

[Package plyr version 1.8.9 Index]