rbind.fill {plyr}	R Documentation

Combine data.frames by row, filling in missing columns.

Description

rbinds a list of data frames filling missing columns with NA.

Usage

  rbind.fill(...)

Arguments

...
input data frames to row bind together. The first argument can be a list of data frames, in which case all other arguments are ignored.

Details

This is an enhancement to rbind that adds in columns that are not present in all inputs, accepts a list of data frames, and operates substantially faster.

Column names and types in the output will appear in the order in which they were encountered. No checking is performed to ensure that each column is of consistent type in the inputs.

Value

a single data frame

See Also

Other binding functions: rbind.fill.matrix

Examples

rbind.fill(mtcars[c("mpg", "wt")], mtcars[c("wt", "cyl")])

[Package plyr version 1.8 Index]