rdply {plyr}	R Documentation

Replicate expression and return results in a data frame.

Description

Evalulate expression n times then combine results into a data frame

Usage

  rdply(.n, .expr, .progress = "none")

Arguments

.n
number of times to evaluate the expression

.expr
expression to evaluate

.progress
name of the progress bar to use, see create_progress_bar

Details

This function runs an expression multiple times, and combines the result into a data frame. If there are no results, then this function returns a data frame with zero rows and columns (data.frame()). This function is equivalent to replicate, but will always return results as a data frame.

Value

a data frame

References

Hadley Wickham (2011). The Split-Apply-Combine Strategy for Data Analysis. Journal of Statistical Software, 40(1), 1-29. http://www.jstatsoft.org/v40/i01/.

Examples

rdply(20, mean(runif(100)))
rdply(20, each(mean, var)(runif(100)))
rdply(20, data.frame(x = runif(2)))

[Package plyr version 1.8 Index]