split_indices {plyr}	R Documentation

Split indices.

Description

An optimised version of split for the special case of splitting row indices into groups, as used by splitter_d

Usage

  split_indices(group, n = max(group))

Arguments

index
integer indices

group
integer groups

n
largest integer (may not appear in index)

[Package plyr version 1.8 Index]