strip_splits {plyr}	R Documentation

Remove splitting variables from a data frame.

Description

This is useful when you want to perform some operation to every column in the data frame, except the variables that you have used to split it. These variables will be automatically added back on to the result when combining all results together.

Usage

strip_splits(df)

Arguments

df
data frame produced by d*ply.

Examples

dlply(mtcars, c("vs", "am"))
dlply(mtcars, c("vs", "am"), strip_splits)

[Package plyr version 1.8.7 Index]