try_default {plyr}	R Documentation

Try, with default in case of error.

Description

try_default wraps try so that it returns a default value in the case of error. tryNULL provides a useful special case when dealing with lists.

Usage

try_default(expr, default, quiet = FALSE)

tryNULL(expr)

Arguments

expr
expression to try

default
default value in case of error

quiet
should errors be printed (TRUE) or ignored (FALSE, default)

See Also

tryapply

[Package plyr version 1.8.7 Index]