Split array, apply function, and return results in a list.

For each slice of an array, apply function then combine results into a list.

alply(.data, .margins, .fun = NULL, ..., .expand = TRUE,
  .progress = "none", .inform = FALSE, .parallel = FALSE,
  .paropts = NULL, .dims = FALSE)

Arguments

.data
matrix, array or data frame to be processed

.margins
a vector giving the subscripts to split up data by. 1 splits up by rows, 2 by columns and c(1,2) by rows and columns, and so on for higher dimensions

.fun
function to apply to each piece

...
other arguments passed on to .fun

.expand
if .data is a data frame, should output be 1d (expand = FALSE), with an element for each row; or nd (expand = TRUE), with a dimension for each variable.

.progress
name of the progress bar to use, see create_progress_bar

.inform
produce informative error messages? This is turned off by default because it substantially slows processing speed, but is very useful for debugging

.parallel
if TRUE, apply function in parallel, using parallel backend provided by foreach

.paropts
a list of additional options passed into the foreach function when parallel computation is enabled. This is important if (for example) your code relies on external data or packages: use the .export and .packages arguments to supply them so that all cluster nodes have the correct environment set up for computing.

.dims
if TRUE, copy over dimensions and names from input.

Value

list of results

Details

The list will have "dims" and "dimnames" corresponding to the margins given. For instance alply(x, c(3,2), ...) where x has dims c(4,3,2) will give a result with dims c(2,3).

alply is somewhat similar to apply for cases where the results are not atomic.

Input

This function splits matrices, arrays and data frames by dimensions

Output

If there are no results, then this function will return a list of length 0 (list()).

References

Hadley Wickham (2011). The Split-Apply-Combine Strategy for Data Analysis. Journal of Statistical Software, 40(1), 1-29. http://www.jstatsoft.org/v40/i01/.

See also

Examples

alply(ozone, 3, quantile)

#> $`1`
#>   0%  25%  50%  75% 100% 
#>  242  248  252  260  312 
#> 
#> $`2`
#>   0%  25%  50%  75% 100% 
#>  238  248  252  258  334 
#> 
#> $`3`
#>    0%   25%   50%   75%  100% 
#> 240.0 250.0 254.0 260.5 338.0 
#> 
#> $`4`
#>   0%  25%  50%  75% 100% 
#>  242  248  252  258  334 
#> 
#> $`5`
#>   0%  25%  50%  75% 100% 
#>  242  250  256  266  348 
#> 
#> $`6`
#>   0%  25%  50%  75% 100% 
#>  246  256  260  278  330 
#> 
#> $`7`
#>   0%  25%  50%  75% 100% 
#>  250  262  266  282  302 
#> 
#> $`8`
#>   0%  25%  50%  75% 100% 
#>  256  268  270  280  292 
#> 
#> $`9`
#>   0%  25%  50%  75% 100% 
#>  258  266  270  272  290 
#> 
#> $`10`
#>   0%  25%  50%  75% 100% 
#>  248  262  266  270  284 
#> 
#> $`11`
#>    0%   25%   50%   75%  100% 
#> 246.0 253.5 258.0 268.0 304.0 
#> 
#> $`12`
#>   0%  25%  50%  75% 100% 
#>  240  248  254  266  314 
#> 
#> $`13`
#>   0%  25%  50%  75% 100% 
#>  238  252  258  268  318 
#> 
#> $`14`
#>    0%   25%   50%   75%  100% 
#> 246.0 252.0 260.0 274.5 344.0 
#> 
#> $`15`
#>   0%  25%  50%  75% 100% 
#>  242  250  254  272  356 
#> 
#> $`16`
#>    0%   25%   50%   75%  100% 
#> 240.0 250.0 256.0 278.5 352.0 
#> 
#> $`17`
#>   0%  25%  50%  75% 100% 
#>  238  250  258  278  324 
#> 
#> $`18`
#>   0%  25%  50%  75% 100% 
#>  244  254  262  284  316 
#> 
#> $`19`
#>   0%  25%  50%  75% 100% 
#>  250  258  268  284  302 
#> 
#> $`20`
#>   0%  25%  50%  75% 100% 
#>  254  264  270  282  300 
#> 
#> $`21`
#>   0%  25%  50%  75% 100% 
#>  256  264  268  274  292 
#> 
#> $`22`
#>   0%  25%  50%  75% 100% 
#>  252  260  264  270  300 
#> 
#> $`23`
#>   0%  25%  50%  75% 100% 
#>  246  252  258  266  294 
#> 
#> $`24`
#>   0%  25%  50%  75% 100% 
#>  244  250  258  266  302 
#> 
#> $`25`
#>   0%  25%  50%  75% 100% 
#>  238  248  254  262  322 
#> 
#> $`26`
#>   0%  25%  50%  75% 100% 
#>  242  252  256  264  344 
#> 
#> $`27`
#>   0%  25%  50%  75% 100% 
#>  236  246  252  262  306 
#> 
#> $`28`
#>   0%  25%  50%  75% 100% 
#>  244  252  255  274  358 
#> 
#> $`29`
#>    0%   25%   50%   75%  100% 
#> 240.0 250.0 256.0 280.5 330.0 
#> 
#> $`30`
#>    0%   25%   50%   75%  100% 
#> 246.0 256.0 260.0 278.5 318.0 
#> 
#> $`31`
#>   0%  25%  50%  75% 100% 
#>  254  264  270  290  314 
#> 
#> $`32`
#>   0%  25%  50%  75% 100% 
#>  252  264  274  282  298 
#> 
#> $`33`
#>   0%  25%  50%  75% 100% 
#>  252  264  270  274  292 
#> 
#> $`34`
#>   0%  25%  50%  75% 100% 
#>  252  260  264  272  304 
#> 
#> $`35`
#>   0%  25%  50%  75% 100% 
#>  242  252  256  266  308 
#> 
#> $`36`
#>   0%  25%  50%  75% 100% 
#>  240  248  252  264  324 
#> 
#> $`37`
#>   0%  25%  50%  75% 100% 
#>  232  242  248  266  334 
#> 
#> $`38`
#>   0%  25%  50%  75% 100% 
#>  234  244  248  264  354 
#> 
#> $`39`
#>   0%  25%  50%  75% 100% 
#>  238  246  250  274  364 
#> 
#> $`40`
#>   0%  25%  50%  75% 100% 
#>  238  244  250  278  370 
#> 
#> $`41`
#>   0%  25%  50%  75% 100% 
#>  242  250  260  284  366 
#> 
#> $`42`
#>   0%  25%  50%  75% 100% 
#>  242  254  262  284  338 
#> 
#> $`43`
#>   0%  25%  50%  75% 100% 
#>  250  262  272  292  308 
#> 
#> $`44`
#>   0%  25%  50%  75% 100% 
#>  250  266  272  280  294 
#> 
#> $`45`
#>   0%  25%  50%  75% 100% 
#>  256  266  270  276  296 
#> 
#> $`46`
#>   0%  25%  50%  75% 100% 
#>  250  260  266  274  314 
#> 
#> $`47`
#>    0%   25%   50%   75%  100% 
#> 242.0 252.0 260.0 270.5 306.0 
#> 
#> $`48`
#>   0%  25%  50%  75% 100% 
#>  244  252  258  266  322 
#> 
#> $`49`
#>   0%  25%  50%  75% 100% 
#>  240  252  256  264  320 
#> 
#> $`50`
#>   0%  25%  50%  75% 100% 
#>  236  250  256  270  342 
#> 
#> $`51`
#>    0%   25%   50%   75%  100% 
#> 240.0 250.0 256.0 272.5 366.0 
#> 
#> $`52`
#>   0%  25%  50%  75% 100% 
#>  238  250  256  274  390 
#> 
#> $`53`
#>   0%  25%  50%  75% 100% 
#>  242  254  260  284  362 
#> 
#> $`54`
#>   0%  25%  50%  75% 100% 
#>  244  260  266  282  334 
#> 
#> $`55`
#>   0%  25%  50%  75% 100% 
#>  250  268  274  288  304 
#> 
#> $`56`
#>   0%  25%  50%  75% 100% 
#>  250  270  275  282  298 
#> 
#> $`57`
#>   0%  25%  50%  75% 100% 
#>  252  270  272  278  296 
#> 
#> $`58`
#>   0%  25%  50%  75% 100% 
#>  254  264  270  276  296 
#> 
#> $`59`
#>   0%  25%  50%  75% 100% 
#>  250  258  264  274  296 
#> 
#> $`60`
#>   0%  25%  50%  75% 100% 
#>  244  256  262  274  328 
#> 
#> $`61`
#>   0%  25%  50%  75% 100% 
#>  244  256  264  272  324 
#> 
#> $`62`
#>   0%  25%  50%  75% 100% 
#>  242  254  262  278  366 
#> 
#> $`63`
#>   0%  25%  50%  75% 100% 
#>  242  252  260  278  368 
#> 
#> $`64`
#>   0%  25%  50%  75% 100% 
#>  240  252  260  282  350 
#> 
#> $`65`
#>   0%  25%  50%  75% 100% 
#>  242  252  262  286  338 
#> 
#> $`66`
#>    0%   25%   50%   75%  100% 
#> 244.0 258.0 270.0 282.5 308.0 
#> 
#> $`67`
#>   0%  25%  50%  75% 100% 
#>  250  264  274  290  306 
#> 
#> $`68`
#>   0%  25%  50%  75% 100% 
#>  250  268  274  286  312 
#> 
#> $`69`
#>   0%  25%  50%  75% 100% 
#>  256  264  268  274  306 
#> 
#> $`70`
#>   0%  25%  50%  75% 100% 
#>  246  256  262  272  312 
#> 
#> $`71`
#>   0%  25%  50%  75% 100% 
#>  240  248  254  268  320 
#> 
#> $`72`
#>   0%  25%  50%  75% 100% 
#>  238  248  254  266  328 
#> 
#> attr(,"split_type")
#> [1] "array"
#> attr(,"split_labels")
#>    time
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> 7     7
#> 8     8
#> 9     9
#> 10   10
#> 11   11
#> 12   12
#> 13   13
#> 14   14
#> 15   15
#> 16   16
#> 17   17
#> 18   18
#> 19   19
#> 20   20
#> 21   21
#> 22   22
#> 23   23
#> 24   24
#> 25   25
#> 26   26
#> 27   27
#> 28   28
#> 29   29
#> 30   30
#> 31   31
#> 32   32
#> 33   33
#> 34   34
#> 35   35
#> 36   36
#> 37   37
#> 38   38
#> 39   39
#> 40   40
#> 41   41
#> 42   42
#> 43   43
#> 44   44
#> 45   45
#> 46   46
#> 47   47
#> 48   48
#> 49   49
#> 50   50
#> 51   51
#> 52   52
#> 53   53
#> 54   54
#> 55   55
#> 56   56
#> 57   57
#> 58   58
#> 59   59
#> 60   60
#> 61   61
#> 62   62
#> 63   63
#> 64   64
#> 65   65
#> 66   66
#> 67   67
#> 68   68
#> 69   69
#> 70   70
#> 71   71
#> 72   72

alply(ozone, 3, function(x) table(round(x)))

#> $`1`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   7  21  65  89  66  55  50  35  37  16  11  10   8   4   9   6   6   3   1   5 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 
#>   3   4   5   6   3   2   8   2  10   3   2   3   6   3   9   3 
#> 
#> $`2`
#> 
#> 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 
#>   1   4  15  28  81  74  68  72  54  31  19   9   6   4   7  11   2   4   6   5 
#> 278 280 282 284 286 288 290 292 294 296 298 302 304 306 308 310 312 314 316 320 
#>   4   2   2   8   4  11   5   6   1   2   3   4   1   1   2   2   4   2   2   1 
#> 322 324 328 330 332 334 
#>   1   2   1   2   1   1 
#> 
#> $`3`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   1   1   1  22  64  78  67  73  56  51  18  11  11   2   2   7  11   4   2   5 
#> 280 282 284 286 288 290 292 294 296 298 300 302 304 306 310 312 314 322 324 326 
#>   5   4   9   4   8   6   1   7   4   3   8   1   3   2   2   2   6   1   1   4 
#> 328 330 338 
#>   1   3   4 
#> 
#> $`4`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>  18  47  67  62  81  89  53  12   9   4   6  10   7   2   4   7   5   6   2   3 
#> 282 284 286 288 290 292 294 296 298 300 302 306 308 310 312 314 316 318 320 322 
#>   6   3   8   3   1   3   3   1   3   3   7   9   3   1   3   1   2   2   4   7 
#> 324 326 328 330 332 334 
#>   1   1   3   1   1   2 
#> 
#> $`5`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>  12  33  45  34  50  36  57  41  36  27  13  26  23   3   7   9   6   8   3  10 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 
#>   4   5  11   4   5   2   9   5   3   3   6   6   2   8   2   2   1   3   3   3 
#> 326 330 332 344 348 
#>   4   1   2   1   2 
#> 
#> $`6`
#> 
#> 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>   1   2   3  12  69  70  75  67  38  17  15  18  13  11  11   6   7   8  13   9 
#> 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 322 324 
#>   8   5  10   4   7   2  12   8   6   9   2  12   4   2   2   4   4   3   1   1 
#> 326 328 330 
#>   1   1   3 
#> 
#> $`7`
#> 
#> 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 
#>   1   2   5   7  20  34  83  92  48  22  23  18  19  16  17  15  16  22  22  17 
#> 290 292 294 296 298 300 302 
#>  17  19  17  11   7   2   4 
#> 
#> $`8`
#> 
#> 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 
#>   1   2   1   8  36  75  91  77  57  29  27  19  30  47  26  33   6   6   5 
#> 
#> $`9`
#> 
#> 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 290 
#>   1  12  27  41  89 104  91  83  59  27  13   6   8   7   5   3 
#> 
#> $`10`
#> 
#> 248 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>   1   2   5  11  32  47  65  78  88  62  71  45  26  21  10   5   3   4 
#> 
#> $`11`
#> 
#> 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>   2  19  50  73  71  60  44  29  22  19  28  18  15  15   7  11   9  12  11  19 
#> 286 288 290 292 294 296 298 300 302 304 
#>   7   6   5   5   2   7   3   1   5   1 
#> 
#> $`12`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   3   2  16  49  96  52  51  40  28  22  25  26  13  12  24  11  13  11   1   3 
#> 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 
#>   7   9   6   1   7   3   9   6   8   6   1   2   4   1   2   1   3   2 
#> 
#> $`13`
#> 
#> 238 240 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   1   2   2  11  33  49  65  75  38  58  30  29  25  13  10  11  13   9   4  14 
#> 280 282 284 286 288 290 292 294 296 298 302 306 308 310 312 316 318 
#>   7   6   4   6   6   7   5   6   3   5   4   5   4   2   8   4   2 
#> 
#> $`14`
#> 
#> 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>   8  30  58  58  37  45  39  47  41  21  10   8  11   8  11  11   4   6   5  12 
#> 286 288 290 292 294 296 298 300 302 304 306 310 312 314 316 318 320 322 324 326 
#>   7   3   5   4   4   7   4   8   5   2   6   6   5   4   5   5   4   4   3   3 
#> 328 332 336 338 340 344 
#>   3   2   2   2   1   2 
#> 
#> $`15`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   3   2  25  58  82  64  56  41  22  15  10  16  11   8  15   8  11   5   8   7 
#> 282 284 286 288 290 292 296 298 300 302 304 306 308 310 312 314 316 318 320 322 
#>   6   4   8   6   1   6   4   3   2   2   2   1   7   1   4   1   1   1   4   5 
#> 324 326 328 330 332 334 336 338 340 342 344 346 348 350 354 356 
#>   4   2   5   2   1   1   4   3   5   4   1   1   2   1   3   1 
#> 
#> $`16`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   5   7  14  29  71  63  45  46  29  33  13  13  11  11  11   7   7   8   3   6 
#> 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 
#>   7   2   4   4   6   8   4   3   7   3   2   4  13   2   2   1   2   6   2   6 
#> 320 322 324 326 328 330 332 334 336 338 340 342 344 346 348 350 352 
#>   2   4   1   2   5   7   4   6   2   3   1   3   6   4   1   4   1 
#> 
#> $`17`
#> 
#> 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 
#>   6   5  17  24  26  45  71  42  22  24  22  15  23  16  19  15   8   6   7  10 
#> 278 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 
#>  14   5  10   5   6  11   9   6   7   8   8   9   4  10   2   5  10   4   5   6 
#> 318 320 322 324 
#>   3   2   2   2 
#> 
#> $`18`
#> 
#> 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 
#>   2  12  23  30  43  83  41  30  15  23  18  19  13  18  12   6   9  14  10   7 
#> 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 
#>   9   8   7   9  12   5  11  17   8  13  12   6   7  10  10   3   1 
#> 
#> $`19`
#> 
#> 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 
#>   2   9  35  49  61  44  23  26  24  18  16  12  22  16  17  17  19  23  19  22 
#> 290 292 294 296 298 300 302 
#>  29  25  23   4  11   4   6 
#> 
#> $`20`
#> 
#> 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 
#>   1   4  10  36  74  85  25  27  37  20  30  31  27  20  21  38  24  24  19  12 
#> 294 296 300 
#>   6   3   2 
#> 
#> $`21`
#> 
#> 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 292 
#>   1   2  20  70  84  87  78  49  36  35  23  31  20  11  14   6   7   2 
#> 
#> $`22`
#> 
#> 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 
#>   1  11  60  65  62  70  57  60  36  31  22  19  19  15  12  13   5   5   4   3 
#> 292 300 
#>   3   3 
#> 
#> $`23`
#> 
#> 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>  17  46  67  50  49  54  32  26  31  32  30  31  14  10  21   9  11   9  11   4 
#> 286 288 290 292 294 
#>   7   6   1   6   2 
#> 
#> $`24`
#> 
#> 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 
#>  17  46  53  53  38  39  32  34  44  37  20  30  11  14  17   9   3  11   7  11 
#> 284 286 288 290 292 294 296 298 300 302 
#>  10   5   6   9   4   4   1   5   3   3 
#> 
#> $`25`
#> 
#> 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 
#>   3   1   9  36  48  61  60  58  39  29  28  27  39  17   7   2  11   4   6   7 
#> 278 280 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 
#>   2   7   5   7   7   5   4   2   3   6   3   2   4   3   4  10   1   2   1   4 
#> 322 
#>   2 
#> 
#> $`26`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   2   9  30  38  64  73  61  44  37  24  27  28  12   4   9  11   5   9  12   5 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 312 314 318 322 326 328 
#>  10   6   7   6   1   1   1   1   2   4   4   4   5   3   1   3   1   2   2   4 
#> 330 332 344 
#>   1   1   2 
#> 
#> $`27`
#> 
#> 236 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 
#>   1   1   8  27  64  70  58  37  28  35  36  35  21  15   8  11   9   7   6  14 
#> 276 278 280 282 284 286 288 290 292 294 296 298 300 302 304 306 
#>   3   9   6   6   6   5   4   4   9   6   4   7   3   4   5   4 
#> 
#> $`28`
#> 
#> 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 
#>   7  13  32  68  86  82  51  19  15  20   6   8   9   9   4   7   5   5   6   5 
#> 284 286 288 290 292 294 296 298 300 302 306 308 310 312 314 316 318 320 322 324 
#>   4   3   5   5   5   3   8   4   1   2   3   3   6   9   2   2   3   2   3   3 
#> 326 328 330 332 334 336 338 340 342 344 346 350 352 354 356 358 
#>   5   6   5   2   3   1   3   3   2   1   4   2   1   2   1   2 
#> 
#> $`29`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   2   4  20  23  68  91  45  26  17  20  29   6  15  16   9   4   5   8   9  11 
#> 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 
#>   4  11   2   9   6   7   8   9   4   8   4   5  14   2   4   3   2   5  12   8 
#> 320 324 326 328 330 
#>   3   2   5   9   2 
#> 
#> $`30`
#> 
#> 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>   1   3   6  31  38  85  76  55  47  22  11   9   7  13   9  11   8   7   6   8 
#> 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 
#>   9   9   5  12   6  12   8   6   7   9   9   4   5   5  11   4   2 
#> 
#> $`31`
#> 
#> 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 
#>   1   9  27  36  57  68  37  33  28  18  17  15  10  12  11  11  12  22  28  27 
#> 294 296 298 300 302 304 306 308 310 314 
#>  28  27  12   9   3   9   2   4   1   2 
#> 
#> $`32`
#> 
#> 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 
#>   2   2   3  13  23  44  67  36  35  28  26  18  19  38  56  43  42  24  22  10 
#> 292 294 296 298 
#>   6  10   7   2 
#> 
#> $`33`
#> 
#> 252 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 
#>   2   3   3  11  45  87  67  48  57  83  79  34  14  13   8   4   6   4   7   1 
#> 
#> $`34`
#> 
#> 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 
#>   7  45  41  37  63  92  53  38  22  26  20  17  14  18  13  10   9   3  11  12 
#> 292 294 296 300 304 
#>  10   4   8   1   2 
#> 
#> $`35`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   1   5  32  39  61  52  63  47  27  43  26  28  15  15  15  12  13   4   1   9 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 
#>   8   7   6   5   7   3   7   4   6   6   3   1   1   4 
#> 
#> $`36`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   8  29  44  45  59  62  45  49  48  22  11   8  15   7   4   4   8   3   3   7 
#> 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 
#>   6   2   4   4   1   2   6   3   7   5   4   7   1   1   3   6   4   5   4   2 
#> 320 322 324 
#>   4  11   3 
#> 
#> $`37`
#> 
#> 232 234 236 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 
#>   6   6  36  40  46  38  46  48  38  51  27   7  13  11   7   7   4  10   7   5 
#> 272 274 276 278 280 282 284 286 288 290 292 294 296 298 302 304 306 308 310 312 
#>   3  10   5   2   2   3   4   1   1   5   8   4   1   3  13   5   2   2   4   5 
#> 314 316 318 320 322 324 326 328 330 332 334 
#>   3   6   2   5   4   7   3   4   2   3   1 
#> 
#> $`38`
#> 
#> 234 236 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 
#>   5  13  35  22  37  71  83  60  35  31   8  10  13   4   4  10  11   3   4   9 
#> 274 276 278 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 314 
#>   4   2   2   3   3   6   1   2   8   3   1   3   2   3   2   1   7   1   4   2 
#> 316 318 320 322 324 328 330 332 334 336 338 340 342 348 352 354 
#>   1   2   2   2   4   4   5   5   1   1   2   6   4   6   1   2 
#> 
#> $`39`
#> 
#> 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 
#>   4  19  49  62  81  60  47  19  22  11  11   9   7   4   6  12   6   2   3   3 
#> 278 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 
#>   8   5   5   5   3   9   4   2   3   3   4   7   3   3   1   1   3   3   2   3 
#> 318 320 322 324 326 330 332 334 336 338 340 342 344 346 348 350 352 354 356 358 
#>   2   5   3   2   2   2   2   1   7   3   7   5   1   3   1   2   1   2   1   4 
#> 360 362 364 
#>   2   1   3 
#> 
#> $`40`
#> 
#> 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 
#>   3  30  76  84  54  25  21  13  16  10  17  16  15  12   8   9   8   4   5   2 
#> 278 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 
#>   8   5   3   5   3  10   2   3   2   4   4   4   4   5   1   4   1   4   6   3 
#> 318 320 324 326 330 332 334 336 338 340 342 344 348 350 352 354 356 358 360 362 
#>   4   5   1   1   3   3   4   3   3   3   6   4   2   2   4   5   2   4   1   1 
#> 364 368 370 
#>   2   2   2 
#> 
#> $`41`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   7  24  41  61  39  31  31  25  21  21  17  14  10   7  15  10  13   5  18   4 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 
#>   9  15   5   4   5   3   1   8   5   5   8   4   3   4   4   2  10   1   4   2 
#> 322 324 326 328 330 332 334 336 338 340 342 344 348 350 354 356 362 364 366 
#>   1   4   6   7   3   2   3   1   5   3   3   1   4   4   3   2   3   4   1 
#> 
#> $`42`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   1   8   7  16  29  54  78  38  20  15  25   8  17  20   7  10  15  17   8   9 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 
#>  11  21  10   7  13   7   7   5   5   7  11   5  13   9   9   3   6   4   1   1 
#> 322 324 328 330 334 336 338 
#>   3   4   1   1   2   1   7 
#> 
#> $`43`
#> 
#> 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 
#>   3   2   4  14  28  41  68  39  27  32  15  18  13  14  21  14  10  12  12  18 
#> 290 292 294 296 298 300 302 304 306 308 
#>  23  21  33  50  24   4  10   1   3   2 
#> 
#> $`44`
#> 
#> 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 
#>   2   2   2   1   4  24  30  49  52  60  40  26  27  37  42  35  27  26  28  29 
#> 290 292 294 
#>  16   9   8 
#> 
#> $`45`
#> 
#> 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 294 296 
#>   4  24  26  40  37  51  55  85  61  42  45  34  16  16  16   7  12   1   1   3 
#> 
#> $`46`
#> 
#> 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 
#>   4  13  39  29  44  50  44  48  59  51  32  14  27  12  22  12  10  19  15   9 
#> 290 292 294 296 298 306 310 314 
#>   7   4   3   2   1   2   1   3 
#> 
#> $`47`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   9  16  21  40  47  53  37  29  30  29  23  25  23  24  26  26  18  16  10  22 
#> 282 284 286 288 290 292 294 296 298 300 304 306 
#>   9   9   7   8   3   5   4   1   2   1   1   2 
#> 
#> $`48`
#> 
#> 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 
#>   3  16  38  63  49  59  36  37  46  36  29  30  29  17  17  15   5   6   2   7 
#> 284 286 288 290 292 294 296 298 300 308 312 316 320 322 
#>   7   5   3   4   2   1   2   3   1   2   2   1   2   1 
#> 
#> $`49`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   4  16  32  20  26  28  47  59  61  31  34  51  27  13   9   8   7   6   4   7 
#> 280 282 284 286 288 290 292 294 296 298 300 302 306 308 310 312 314 316 318 320 
#>   5   7   7   9   3   3   2   7   4   6   4   2   5   2   8   4   1   2   3   2 
#> 
#> $`50`
#> 
#> 236 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 
#>   1   2   7  33  29  37  46  48  45  41  40  41  40   5   9   6   9   8   8  11 
#> 278 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 314 316 318 
#>   3   3   7   4   5   1   4   2   2   5   8   2   3   4   1   5   4   1   4   6 
#> 320 322 324 326 328 332 334 336 338 340 342 
#>   2   3   1   5   4   7   2   1   3   4   4 
#> 
#> $`51`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   1   5  34  38  45  42  42  52  66  41  13  11  12   8  10   8   4   8   8   7 
#> 280 282 284 286 288 292 294 296 298 300 302 304 306 308 312 314 316 318 320 322 
#>   4   1   6   6   3   4   5   9   3   3   5   1   1   2   1   5   6  10   4   1 
#> 324 326 328 330 332 336 338 340 342 346 348 352 354 358 360 362 366 
#>   2   2   2   3   6   5   2   1   2   1   2   2   3   4   1   1   2 
#> 
#> $`52`
#> 
#> 238 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   1  12  22  24  33  60  54  48  43  19  16  15  19  15  19   9  17   9   8   7 
#> 280 282 284 286 288 290 292 294 296 298 300 304 306 308 310 312 314 316 318 320 
#>   7   7   6   5   4   3   6   6   1   3   6   1   5   2   4   3   2   2   1   3 
#> 322 324 326 328 330 334 338 340 342 344 350 354 356 360 364 366 390 
#>   1   7   3   6   4   4   4   1   4   1   2   3   1   2   1   4   1 
#> 
#> $`53`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   5   2  11  26  25  17  59  47  43  57  21  12  11  12  21  22   9   4  15   2 
#> 282 284 286 288 290 294 296 298 300 302 304 306 308 310 312 314 316 318 320 322 
#>   5   8   7   4   4   9   5   4   6   2   2   6   7   6   6   1   5   4   9   4 
#> 324 326 328 330 332 334 336 338 340 342 344 346 348 350 362 
#>   2   1   7   7   6   5   4   2   5   2   4   1   3   1   1 
#> 
#> $`54`
#> 
#> 244 246 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 
#>   3   1  14  24  40  32  51  71  42  31  26   8   7  22  13  12  22  14  12   8 
#> 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 322 324 326 
#>   9   6  13   4   4   3   8   8   4   6   8  14   7   7   1   5   2   4   4   3 
#> 330 334 
#>   1   2 
#> 
#> $`55`
#> 
#> 250 252 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 294 296 
#>   4   1   2  33  69  55  51  67  35  23  15  11  15  17  31  26  41  24  26  13 
#> 298 300 302 304 
#>   7   5   3   2 
#> 
#> $`56`
#> 
#> 250 252 254 258 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 
#>   3   1   1   1   1   7  14  59  80  66  55  52  37  51  37  15  12  12  26  24 
#> 294 296 298 
#>  11   8   3 
#> 
#> $`57`
#> 
#> 252 254 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 
#>   1   5   6   2   7  23  30  45  95  87  61  55  43  40  18  12  21  10   3   6 
#> 294 296 
#>   3   3 
#> 
#> $`58`
#> 
#> 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 292 294 
#>   1   2  30  54  55  53  33  33  32  76  57  29  29  21  19   6  14  19  10   2 
#> 296 
#>   1 
#> 
#> $`59`
#> 
#> 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 
#>   1  11  51  66  72  34  30  34  35  18  16  32  36  20  26  22  26  11  14  14 
#> 290 292 294 296 
#>   3   2   1   1 
#> 
#> $`60`
#> 
#> 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 
#>   3  29  38  25  23  22  34  36  44  40  46  29  25  17  17  24  11  12  10  10 
#> 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 322 
#>   7   4  13   3   2  11   6   2   2   4   2   2   3   2   1   2   7   1   2   1 
#> 324 326 328 
#>   1   2   1 
#> 
#> $`61`
#> 
#> 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 
#>   6   9  19  25  29  40  23  36  48  45  39  43  33  32  17   6  11  14   7   3 
#> 284 286 288 290 292 294 296 298 300 302 304 306 310 312 316 318 320 322 324 
#>   9   8  10   3   7   2   1   3   6   9   3   7   5   1   4   4   4   3   2 
#> 
#> $`62`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   1   9   7  16  37  48  45  32  25  28  46  49  16  11  17  22  11   7  14   7 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 
#>   9  10   3   6   8   7   1   1   6   4   6   3   5   3   6   7   5   6   3   1 
#> 322 324 326 328 330 332 334 336 338 340 366 
#>   3   2   2   8   3   2   3   1   2   1   1 
#> 
#> $`63`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   1   5  12  32  46  57  36  35  40  37  25  11   9  13  13  25  12  15  10  12 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 
#>   5   6   5   9   2   3   4   6   4   6   5   5   2   2   1   7   2   3   8   5 
#> 324 326 328 330 332 334 336 338 340 342 344 346 348 350 352 368 
#>   3   1   3   2   3   5   3   3   4   3   1   1   1   3   2   2 
#> 
#> $`64`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   3  23  21  24  32  35  47  30  29  33  30  15   8   8  15   8  25  15  12   8 
#> 280 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 320 
#>   6  14   6   8   3  12   6   2   6   5   6   4   8   5   3   2   2   9   7   7 
#> 322 324 326 328 330 332 334 336 338 340 342 344 350 
#>   7   5   3   1   2   2   1   3   1   4   3   1   1 
#> 
#> $`65`
#> 
#> 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 
#>   5  24  16  28  39  54  52  28  17  20  26  18  20  16  12  11  10   7   5   2 
#> 282 284 286 288 290 292 294 296 298 300 302 304 306 308 310 312 314 316 318 320 
#>  10   5   8   3  11   8   5   9  10   3   3   8   7   6   5   4   6   9   5   7 
#> 322 324 326 328 330 332 334 338 
#>   5   5   3   6   4   4   5   2 
#> 
#> $`66`
#> 
#> 244 246 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>   3   1   6  29  33  37  42  48  35  20  12  21  26  29  27  18  19   9  17  14 
#> 286 288 290 292 294 296 298 300 302 304 306 308 
#>   8  17  13  17   9   3  13   7  11  17  13   2 
#> 
#> $`67`
#> 
#> 250 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 
#>   4   1  27  23  25  34  68  27  26  25   8  32  23  17  16  23  15  16  19  25 
#> 292 294 296 298 300 302 304 306 
#>  29  32  26  21   4   4   3   3 
#> 
#> $`68`
#> 
#> 250 254 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 290 292 294 
#>   3   1  12  23  38  54  69  58  23  35  17  20  20  29  27  13  21  20  32  32 
#> 296 298 300 312 
#>  22   2   4   1 
#> 
#> $`69`
#> 
#> 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 286 288 306 
#>   3  22  34  49  40  69  96  51  35  39  32  20  24  22  11  18  10   1 
#> 
#> $`70`
#> 
#> 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 280 282 284 
#>   5  17  25  29  45  42  47  49  35  33  32  31  28  29  29  12  20  23  14   7 
#> 286 288 290 292 294 298 308 312 
#>   8   3   4   3   1   2   1   2 
#> 
#> $`71`
#> 
#> 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 278 
#>   5  20  63  46  43  51  32  37  27  15  23  18  21  21  15  18  16  13  12  12 
#> 280 282 284 286 288 290 292 294 296 298 300 302 314 320 
#>   8   8  12   7   3   6   5   8   2   2   1   2   1   3 
#> 
#> $`72`
#> 
#> 238 240 242 244 246 248 250 252 254 256 258 260 262 264 266 268 270 272 274 276 
#>   2  16  43  22  26  47  53  53  43  36  26  23  24  15  19  15  20  13   4  10 
#> 278 280 282 284 286 288 290 294 296 300 302 304 308 312 314 318 322 324 326 328 
#>   2   9   7   4   4   7   4   4   1   5   2   4   2   2   2   1   1   3   1   1 
#> 
#> attr(,"split_type")
#> [1] "array"
#> attr(,"split_labels")
#>    time
#> 1     1
#> 2     2
#> 3     3
#> 4     4
#> 5     5
#> 6     6
#> 7     7
#> 8     8
#> 9     9
#> 10   10
#> 11   11
#> 12   12
#> 13   13
#> 14   14
#> 15   15
#> 16   16
#> 17   17
#> 18   18
#> 19   19
#> 20   20
#> 21   21
#> 22   22
#> 23   23
#> 24   24
#> 25   25
#> 26   26
#> 27   27
#> 28   28
#> 29   29
#> 30   30
#> 31   31
#> 32   32
#> 33   33
#> 34   34
#> 35   35
#> 36   36
#> 37   37
#> 38   38
#> 39   39
#> 40   40
#> 41   41
#> 42   42
#> 43   43
#> 44   44
#> 45   45
#> 46   46
#> 47   47
#> 48   48
#> 49   49
#> 50   50
#> 51   51
#> 52   52
#> 53   53
#> 54   54
#> 55   55
#> 56   56
#> 57   57
#> 58   58
#> 59   59
#> 60   60
#> 61   61
#> 62   62
#> 63   63
#> 64   64
#> 65   65
#> 66   66
#> 67   67
#> 68   68
#> 69   69
#> 70   70
#> 71   71
#> 72   72