Create progress bar.

Create progress bar object from text string.

create_progress_bar(name = "none", ...)

Arguments

name
type of progress bar to create

...
other arguments passed onto progress bar function

Details

Progress bars give feedback on how apply step is proceeding. This is mainly useful for long running functions, as for short functions, the time taken up by splitting and combining may be on the same order (or longer) as the apply step. Additionally, for short functions, the time needed to update the progress bar can significantly slow down the process. For the trivial examples below, using the tk progress bar slows things down by a factor of a thousand.

Note the that progress bar is approximate, and if the time taken by individual function applications is highly non-uniform it may not be very informative of the time left.

There are currently four types of progress bar: "none", "text", "tk", and "win". See the individual documentation for more details. In plyr functions, these can either be specified by name, or you can create the progress bar object yourself if you want more control over its apperance. See the examples.

See also

Examples

# No progress bar
l_ply(1:100, identity, .progress = "none")
if (FALSE) {
# Use the Tcl/Tk interface
l_ply(1:100, identity, .progress = "tk")
}
# Text-based progress (|======|)
l_ply(1:100, identity, .progress = "text")

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |==========                                                            |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |===============                                                       |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |====================                                                  |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |======================                                                |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=============================                                         |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |================================                                      |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |==================================                                    |  49%  |                                                                              |===================================                                   |  50%  |                                                                              |====================================                                  |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |==============================================                        |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%

# Choose a progress character, run a length of time you can see
l_ply(1:10000, identity, .progress = progress_text(char = "."))

#>   |                                                                              |                                                                      |   0%  |                                                                              |                                                                      |   1%  |                                                                              |.                                                                     |   1%  |                                                                              |.                                                                     |   2%  |                                                                              |..                                                                    |   2%  |                                                                              |..                                                                    |   3%  |                                                                              |..                                                                    |   4%  |                                                                              |...                                                                   |   4%  |                                                                              |...                                                                   |   5%  |                                                                              |....                                                                  |   5%  |                                                                              |....                                                                  |   6%  |                                                                              |.....                                                                 |   6%  |                                                                              |.....                                                                 |   7%  |                                                                              |.....                                                                 |   8%  |                                                                              |......                                                                |   8%  |                                                                              |......                                                                |   9%  |                                                                              |.......                                                               |   9%  |                                                                              |.......                                                               |  10%  |                                                                              |.......                                                               |  11%  |                                                                              |........                                                              |  11%  |                                                                              |........                                                              |  12%  |                                                                              |.........                                                             |  12%  |                                                                              |.........                                                             |  13%  |                                                                              |.........                                                             |  14%  |                                                                              |..........                                                            |  14%  |                                                                              |..........                                                            |  15%  |                                                                              |...........                                                           |  15%  |                                                                              |...........                                                           |  16%  |                                                                              |............                                                          |  16%  |                                                                              |............                                                          |  17%  |                                                                              |............                                                          |  18%  |                                                                              |.............                                                         |  18%  |                                                                              |.............                                                         |  19%  |                                                                              |..............                                                        |  19%  |                                                                              |..............                                                        |  20%  |                                                                              |..............                                                        |  21%  |                                                                              |...............                                                       |  21%  |                                                                              |...............                                                       |  22%  |                                                                              |................                                                      |  22%  |                                                                              |................                                                      |  23%  |                                                                              |................                                                      |  24%  |                                                                              |.................                                                     |  24%  |                                                                              |.................                                                     |  25%  |                                                                              |..................                                                    |  25%  |                                                                              |..................                                                    |  26%  |                                                                              |...................                                                   |  26%  |                                                                              |...................                                                   |  27%  |                                                                              |...................                                                   |  28%  |                                                                              |....................                                                  |  28%  |                                                                              |....................                                                  |  29%  |                                                                              |.....................                                                 |  29%  |                                                                              |.....................                                                 |  30%  |                                                                              |.....................                                                 |  31%  |                                                                              |......................                                                |  31%  |                                                                              |......................                                                |  32%  |                                                                              |.......................                                               |  32%  |                                                                              |.......................                                               |  33%  |                                                                              |.......................                                               |  34%  |                                                                              |........................                                              |  34%  |                                                                              |........................                                              |  35%  |                                                                              |.........................                                             |  35%  |                                                                              |.........................                                             |  36%  |                                                                              |..........................                                            |  36%  |                                                                              |..........................                                            |  37%  |                                                                              |..........................                                            |  38%  |                                                                              |...........................                                           |  38%  |                                                                              |...........................                                           |  39%  |                                                                              |............................                                          |  39%  |                                                                              |............................                                          |  40%  |                                                                              |............................                                          |  41%  |                                                                              |.............................                                         |  41%  |                                                                              |.............................                                         |  42%  |                                                                              |..............................                                        |  42%  |                                                                              |..............................                                        |  43%  |                                                                              |..............................                                        |  44%  |                                                                              |...............................                                       |  44%  |                                                                              |...............................                                       |  45%  |                                                                              |................................                                      |  45%  |                                                                              |................................                                      |  46%  |                                                                              |.................................                                     |  46%  |                                                                              |.................................                                     |  47%  |                                                                              |.................................                                     |  48%  |                                                                              |..................................                                    |  48%  |                                                                              |..................................                                    |  49%  |                                                                              |...................................                                   |  49%  |                                                                              |...................................                                   |  50%  |                                                                              |...................................                                   |  51%  |                                                                              |....................................                                  |  51%  |                                                                              |....................................                                  |  52%  |                                                                              |.....................................                                 |  52%  |                                                                              |.....................................                                 |  53%  |                                                                              |.....................................                                 |  54%  |                                                                              |......................................                                |  54%  |                                                                              |......................................                                |  55%  |                                                                              |.......................................                               |  55%  |                                                                              |.......................................                               |  56%  |                                                                              |........................................                              |  56%  |                                                                              |........................................                              |  57%  |                                                                              |........................................                              |  58%  |                                                                              |.........................................                             |  58%  |                                                                              |.........................................                             |  59%  |                                                                              |..........................................                            |  59%  |                                                                              |..........................................                            |  60%  |                                                                              |..........................................                            |  61%  |                                                                              |...........................................                           |  61%  |                                                                              |...........................................                           |  62%  |                                                                              |............................................                          |  62%  |                                                                              |............................................                          |  63%  |                                                                              |............................................                          |  64%  |                                                                              |.............................................                         |  64%  |                                                                              |.............................................                         |  65%  |                                                                              |..............................................                        |  65%  |                                                                              |..............................................                        |  66%  |                                                                              |...............................................                       |  66%  |                                                                              |...............................................                       |  67%  |                                                                              |...............................................                       |  68%  |                                                                              |................................................                      |  68%  |                                                                              |................................................                      |  69%  |                                                                              |.................................................                     |  69%  |                                                                              |.................................................                     |  70%  |                                                                              |.................................................                     |  71%  |                                                                              |..................................................                    |  71%  |                                                                              |..................................................                    |  72%  |                                                                              |...................................................                   |  72%  |                                                                              |...................................................                   |  73%  |                                                                              |...................................................                   |  74%  |                                                                              |....................................................                  |  74%  |                                                                              |....................................................                  |  75%  |                                                                              |.....................................................                 |  75%  |                                                                              |.....................................................                 |  76%  |                                                                              |......................................................                |  76%  |                                                                              |......................................................                |  77%  |                                                                              |......................................................                |  78%  |                                                                              |.......................................................               |  78%  |                                                                              |.......................................................               |  79%  |                                                                              |........................................................              |  79%  |                                                                              |........................................................              |  80%  |                                                                              |........................................................              |  81%  |                                                                              |.........................................................             |  81%  |                                                                              |.........................................................             |  82%  |                                                                              |..........................................................            |  82%  |                                                                              |..........................................................            |  83%  |                                                                              |..........................................................            |  84%  |                                                                              |...........................................................           |  84%  |                                                                              |...........................................................           |  85%  |                                                                              |............................................................          |  85%  |                                                                              |............................................................          |  86%  |                                                                              |.............................................................         |  86%  |                                                                              |.............................................................         |  87%  |                                                                              |.............................................................         |  88%  |                                                                              |..............................................................        |  88%  |                                                                              |..............................................................        |  89%  |                                                                              |...............................................................       |  89%  |                                                                              |...............................................................       |  90%  |                                                                              |...............................................................       |  91%  |                                                                              |................................................................      |  91%  |                                                                              |................................................................      |  92%  |                                                                              |.................................................................     |  92%  |                                                                              |.................................................................     |  93%  |                                                                              |.................................................................     |  94%  |                                                                              |..................................................................    |  94%  |                                                                              |..................................................................    |  95%  |                                                                              |...................................................................   |  95%  |                                                                              |...................................................................   |  96%  |                                                                              |....................................................................  |  96%  |                                                                              |....................................................................  |  97%  |                                                                              |....................................................................  |  98%  |                                                                              |..................................................................... |  98%  |                                                                              |..................................................................... |  99%  |                                                                              |......................................................................|  99%  |                                                                              |......................................................................| 100%