list_to_dataframe {plyr}	R Documentation

List to data frame.

Description

Reduce/simplify a list of homogenous objects to a data frame. All NULL entries are removed. Remaining entries must be all atomic or all data frames.

Usage

list_to_dataframe(res, labels = NULL, id_name = NULL, id_as_factor = FALSE)

Arguments

res
list of input data

labels
a data frame of labels, one row for each element of res

idname
the name of the index column, NULL for no index column

See Also

Other list simplification functions: list_to_array(), list_to_vector()

[Package plyr version 1.8.7 Index]