Mutate a data frame by adding new or replacing existing columns.

This function is very similar to transform but it executes the transformations iteratively so that later transformations can use the columns created by earlier transformations. Like transform, unnamed components are silently dropped.

mutate(.data, ...)

Arguments

.data
the data frame to transform

...
named parameters giving definitions of new columns.

Details

Mutate seems to be considerably faster than transform for large data frames.

See also

Examples

# Examples from transform
mutate(airquality, Ozone = -Ozone)

#>     Ozone Solar.R Wind Temp Month Day
#> 1     -41     190  7.4   67     5   1
#> 2     -36     118  8.0   72     5   2
#> 3     -12     149 12.6   74     5   3
#> 4     -18     313 11.5   62     5   4
#> 5      NA      NA 14.3   56     5   5
#> 6     -28      NA 14.9   66     5   6
#> 7     -23     299  8.6   65     5   7
#> 8     -19      99 13.8   59     5   8
#> 9      -8      19 20.1   61     5   9
#> 10     NA     194  8.6   69     5  10
#> 11     -7      NA  6.9   74     5  11
#> 12    -16     256  9.7   69     5  12
#> 13    -11     290  9.2   66     5  13
#> 14    -14     274 10.9   68     5  14
#> 15    -18      65 13.2   58     5  15
#> 16    -14     334 11.5   64     5  16
#> 17    -34     307 12.0   66     5  17
#> 18     -6      78 18.4   57     5  18
#> 19    -30     322 11.5   68     5  19
#> 20    -11      44  9.7   62     5  20
#> 21     -1       8  9.7   59     5  21
#> 22    -11     320 16.6   73     5  22
#> 23     -4      25  9.7   61     5  23
#> 24    -32      92 12.0   61     5  24
#> 25     NA      66 16.6   57     5  25
#> 26     NA     266 14.9   58     5  26
#> 27     NA      NA  8.0   57     5  27
#> 28    -23      13 12.0   67     5  28
#> 29    -45     252 14.9   81     5  29
#> 30   -115     223  5.7   79     5  30
#> 31    -37     279  7.4   76     5  31
#> 32     NA     286  8.6   78     6   1
#> 33     NA     287  9.7   74     6   2
#> 34     NA     242 16.1   67     6   3
#> 35     NA     186  9.2   84     6   4
#> 36     NA     220  8.6   85     6   5
#> 37     NA     264 14.3   79     6   6
#> 38    -29     127  9.7   82     6   7
#> 39     NA     273  6.9   87     6   8
#> 40    -71     291 13.8   90     6   9
#> 41    -39     323 11.5   87     6  10
#> 42     NA     259 10.9   93     6  11
#> 43     NA     250  9.2   92     6  12
#> 44    -23     148  8.0   82     6  13
#> 45     NA     332 13.8   80     6  14
#> 46     NA     322 11.5   79     6  15
#> 47    -21     191 14.9   77     6  16
#> 48    -37     284 20.7   72     6  17
#> 49    -20      37  9.2   65     6  18
#> 50    -12     120 11.5   73     6  19
#> 51    -13     137 10.3   76     6  20
#> 52     NA     150  6.3   77     6  21
#> 53     NA      59  1.7   76     6  22
#> 54     NA      91  4.6   76     6  23
#> 55     NA     250  6.3   76     6  24
#> 56     NA     135  8.0   75     6  25
#> 57     NA     127  8.0   78     6  26
#> 58     NA      47 10.3   73     6  27
#> 59     NA      98 11.5   80     6  28
#> 60     NA      31 14.9   77     6  29
#> 61     NA     138  8.0   83     6  30
#> 62   -135     269  4.1   84     7   1
#> 63    -49     248  9.2   85     7   2
#> 64    -32     236  9.2   81     7   3
#> 65     NA     101 10.9   84     7   4
#> 66    -64     175  4.6   83     7   5
#> 67    -40     314 10.9   83     7   6
#> 68    -77     276  5.1   88     7   7
#> 69    -97     267  6.3   92     7   8
#> 70    -97     272  5.7   92     7   9
#> 71    -85     175  7.4   89     7  10
#> 72     NA     139  8.6   82     7  11
#> 73    -10     264 14.3   73     7  12
#> 74    -27     175 14.9   81     7  13
#> 75     NA     291 14.9   91     7  14
#> 76     -7      48 14.3   80     7  15
#> 77    -48     260  6.9   81     7  16
#> 78    -35     274 10.3   82     7  17
#> 79    -61     285  6.3   84     7  18
#> 80    -79     187  5.1   87     7  19
#> 81    -63     220 11.5   85     7  20
#> 82    -16       7  6.9   74     7  21
#> 83     NA     258  9.7   81     7  22
#> 84     NA     295 11.5   82     7  23
#> 85    -80     294  8.6   86     7  24
#> 86   -108     223  8.0   85     7  25
#> 87    -20      81  8.6   82     7  26
#> 88    -52      82 12.0   86     7  27
#> 89    -82     213  7.4   88     7  28
#> 90    -50     275  7.4   86     7  29
#> 91    -64     253  7.4   83     7  30
#> 92    -59     254  9.2   81     7  31
#> 93    -39      83  6.9   81     8   1
#> 94     -9      24 13.8   81     8   2
#> 95    -16      77  7.4   82     8   3
#> 96    -78      NA  6.9   86     8   4
#> 97    -35      NA  7.4   85     8   5
#> 98    -66      NA  4.6   87     8   6
#> 99   -122     255  4.0   89     8   7
#> 100   -89     229 10.3   90     8   8
#> 101  -110     207  8.0   90     8   9
#> 102    NA     222  8.6   92     8  10
#> 103    NA     137 11.5   86     8  11
#> 104   -44     192 11.5   86     8  12
#> 105   -28     273 11.5   82     8  13
#> 106   -65     157  9.7   80     8  14
#> 107    NA      64 11.5   79     8  15
#> 108   -22      71 10.3   77     8  16
#> 109   -59      51  6.3   79     8  17
#> 110   -23     115  7.4   76     8  18
#> 111   -31     244 10.9   78     8  19
#> 112   -44     190 10.3   78     8  20
#> 113   -21     259 15.5   77     8  21
#> 114    -9      36 14.3   72     8  22
#> 115    NA     255 12.6   75     8  23
#> 116   -45     212  9.7   79     8  24
#> 117  -168     238  3.4   81     8  25
#> 118   -73     215  8.0   86     8  26
#> 119    NA     153  5.7   88     8  27
#> 120   -76     203  9.7   97     8  28
#> 121  -118     225  2.3   94     8  29
#> 122   -84     237  6.3   96     8  30
#> 123   -85     188  6.3   94     8  31
#> 124   -96     167  6.9   91     9   1
#> 125   -78     197  5.1   92     9   2
#> 126   -73     183  2.8   93     9   3
#> 127   -91     189  4.6   93     9   4
#> 128   -47      95  7.4   87     9   5
#> 129   -32      92 15.5   84     9   6
#> 130   -20     252 10.9   80     9   7
#> 131   -23     220 10.3   78     9   8
#> 132   -21     230 10.9   75     9   9
#> 133   -24     259  9.7   73     9  10
#> 134   -44     236 14.9   81     9  11
#> 135   -21     259 15.5   76     9  12
#> 136   -28     238  6.3   77     9  13
#> 137    -9      24 10.9   71     9  14
#> 138   -13     112 11.5   71     9  15
#> 139   -46     237  6.9   78     9  16
#> 140   -18     224 13.8   67     9  17
#> 141   -13      27 10.3   76     9  18
#> 142   -24     238 10.3   68     9  19
#> 143   -16     201  8.0   82     9  20
#> 144   -13     238 12.6   64     9  21
#> 145   -23      14  9.2   71     9  22
#> 146   -36     139 10.3   81     9  23
#> 147    -7      49 10.3   69     9  24
#> 148   -14      20 16.6   63     9  25
#> 149   -30     193  6.9   70     9  26
#> 150    NA     145 13.2   77     9  27
#> 151   -14     191 14.3   75     9  28
#> 152   -18     131  8.0   76     9  29
#> 153   -20     223 11.5   68     9  30

mutate(airquality, new = -Ozone, Temp = (Temp - 32) / 1.8)

#>     Ozone Solar.R Wind     Temp Month Day  new
#> 1      41     190  7.4 19.44444     5   1  -41
#> 2      36     118  8.0 22.22222     5   2  -36
#> 3      12     149 12.6 23.33333     5   3  -12
#> 4      18     313 11.5 16.66667     5   4  -18
#> 5      NA      NA 14.3 13.33333     5   5   NA
#> 6      28      NA 14.9 18.88889     5   6  -28
#> 7      23     299  8.6 18.33333     5   7  -23
#> 8      19      99 13.8 15.00000     5   8  -19
#> 9       8      19 20.1 16.11111     5   9   -8
#> 10     NA     194  8.6 20.55556     5  10   NA
#> 11      7      NA  6.9 23.33333     5  11   -7
#> 12     16     256  9.7 20.55556     5  12  -16
#> 13     11     290  9.2 18.88889     5  13  -11
#> 14     14     274 10.9 20.00000     5  14  -14
#> 15     18      65 13.2 14.44444     5  15  -18
#> 16     14     334 11.5 17.77778     5  16  -14
#> 17     34     307 12.0 18.88889     5  17  -34
#> 18      6      78 18.4 13.88889     5  18   -6
#> 19     30     322 11.5 20.00000     5  19  -30
#> 20     11      44  9.7 16.66667     5  20  -11
#> 21      1       8  9.7 15.00000     5  21   -1
#> 22     11     320 16.6 22.77778     5  22  -11
#> 23      4      25  9.7 16.11111     5  23   -4
#> 24     32      92 12.0 16.11111     5  24  -32
#> 25     NA      66 16.6 13.88889     5  25   NA
#> 26     NA     266 14.9 14.44444     5  26   NA
#> 27     NA      NA  8.0 13.88889     5  27   NA
#> 28     23      13 12.0 19.44444     5  28  -23
#> 29     45     252 14.9 27.22222     5  29  -45
#> 30    115     223  5.7 26.11111     5  30 -115
#> 31     37     279  7.4 24.44444     5  31  -37
#> 32     NA     286  8.6 25.55556     6   1   NA
#> 33     NA     287  9.7 23.33333     6   2   NA
#> 34     NA     242 16.1 19.44444     6   3   NA
#> 35     NA     186  9.2 28.88889     6   4   NA
#> 36     NA     220  8.6 29.44444     6   5   NA
#> 37     NA     264 14.3 26.11111     6   6   NA
#> 38     29     127  9.7 27.77778     6   7  -29
#> 39     NA     273  6.9 30.55556     6   8   NA
#> 40     71     291 13.8 32.22222     6   9  -71
#> 41     39     323 11.5 30.55556     6  10  -39
#> 42     NA     259 10.9 33.88889     6  11   NA
#> 43     NA     250  9.2 33.33333     6  12   NA
#> 44     23     148  8.0 27.77778     6  13  -23
#> 45     NA     332 13.8 26.66667     6  14   NA
#> 46     NA     322 11.5 26.11111     6  15   NA
#> 47     21     191 14.9 25.00000     6  16  -21
#> 48     37     284 20.7 22.22222     6  17  -37
#> 49     20      37  9.2 18.33333     6  18  -20
#> 50     12     120 11.5 22.77778     6  19  -12
#> 51     13     137 10.3 24.44444     6  20  -13
#> 52     NA     150  6.3 25.00000     6  21   NA
#> 53     NA      59  1.7 24.44444     6  22   NA
#> 54     NA      91  4.6 24.44444     6  23   NA
#> 55     NA     250  6.3 24.44444     6  24   NA
#> 56     NA     135  8.0 23.88889     6  25   NA
#> 57     NA     127  8.0 25.55556     6  26   NA
#> 58     NA      47 10.3 22.77778     6  27   NA
#> 59     NA      98 11.5 26.66667     6  28   NA
#> 60     NA      31 14.9 25.00000     6  29   NA
#> 61     NA     138  8.0 28.33333     6  30   NA
#> 62    135     269  4.1 28.88889     7   1 -135
#> 63     49     248  9.2 29.44444     7   2  -49
#> 64     32     236  9.2 27.22222     7   3  -32
#> 65     NA     101 10.9 28.88889     7   4   NA
#> 66     64     175  4.6 28.33333     7   5  -64
#> 67     40     314 10.9 28.33333     7   6  -40
#> 68     77     276  5.1 31.11111     7   7  -77
#> 69     97     267  6.3 33.33333     7   8  -97
#> 70     97     272  5.7 33.33333     7   9  -97
#> 71     85     175  7.4 31.66667     7  10  -85
#> 72     NA     139  8.6 27.77778     7  11   NA
#> 73     10     264 14.3 22.77778     7  12  -10
#> 74     27     175 14.9 27.22222     7  13  -27
#> 75     NA     291 14.9 32.77778     7  14   NA
#> 76      7      48 14.3 26.66667     7  15   -7
#> 77     48     260  6.9 27.22222     7  16  -48
#> 78     35     274 10.3 27.77778     7  17  -35
#> 79     61     285  6.3 28.88889     7  18  -61
#> 80     79     187  5.1 30.55556     7  19  -79
#> 81     63     220 11.5 29.44444     7  20  -63
#> 82     16       7  6.9 23.33333     7  21  -16
#> 83     NA     258  9.7 27.22222     7  22   NA
#> 84     NA     295 11.5 27.77778     7  23   NA
#> 85     80     294  8.6 30.00000     7  24  -80
#> 86    108     223  8.0 29.44444     7  25 -108
#> 87     20      81  8.6 27.77778     7  26  -20
#> 88     52      82 12.0 30.00000     7  27  -52
#> 89     82     213  7.4 31.11111     7  28  -82
#> 90     50     275  7.4 30.00000     7  29  -50
#> 91     64     253  7.4 28.33333     7  30  -64
#> 92     59     254  9.2 27.22222     7  31  -59
#> 93     39      83  6.9 27.22222     8   1  -39
#> 94      9      24 13.8 27.22222     8   2   -9
#> 95     16      77  7.4 27.77778     8   3  -16
#> 96     78      NA  6.9 30.00000     8   4  -78
#> 97     35      NA  7.4 29.44444     8   5  -35
#> 98     66      NA  4.6 30.55556     8   6  -66
#> 99    122     255  4.0 31.66667     8   7 -122
#> 100    89     229 10.3 32.22222     8   8  -89
#> 101   110     207  8.0 32.22222     8   9 -110
#> 102    NA     222  8.6 33.33333     8  10   NA
#> 103    NA     137 11.5 30.00000     8  11   NA
#> 104    44     192 11.5 30.00000     8  12  -44
#> 105    28     273 11.5 27.77778     8  13  -28
#> 106    65     157  9.7 26.66667     8  14  -65
#> 107    NA      64 11.5 26.11111     8  15   NA
#> 108    22      71 10.3 25.00000     8  16  -22
#> 109    59      51  6.3 26.11111     8  17  -59
#> 110    23     115  7.4 24.44444     8  18  -23
#> 111    31     244 10.9 25.55556     8  19  -31
#> 112    44     190 10.3 25.55556     8  20  -44
#> 113    21     259 15.5 25.00000     8  21  -21
#> 114     9      36 14.3 22.22222     8  22   -9
#> 115    NA     255 12.6 23.88889     8  23   NA
#> 116    45     212  9.7 26.11111     8  24  -45
#> 117   168     238  3.4 27.22222     8  25 -168
#> 118    73     215  8.0 30.00000     8  26  -73
#> 119    NA     153  5.7 31.11111     8  27   NA
#> 120    76     203  9.7 36.11111     8  28  -76
#> 121   118     225  2.3 34.44444     8  29 -118
#> 122    84     237  6.3 35.55556     8  30  -84
#> 123    85     188  6.3 34.44444     8  31  -85
#> 124    96     167  6.9 32.77778     9   1  -96
#> 125    78     197  5.1 33.33333     9   2  -78
#> 126    73     183  2.8 33.88889     9   3  -73
#> 127    91     189  4.6 33.88889     9   4  -91
#> 128    47      95  7.4 30.55556     9   5  -47
#> 129    32      92 15.5 28.88889     9   6  -32
#> 130    20     252 10.9 26.66667     9   7  -20
#> 131    23     220 10.3 25.55556     9   8  -23
#> 132    21     230 10.9 23.88889     9   9  -21
#> 133    24     259  9.7 22.77778     9  10  -24
#> 134    44     236 14.9 27.22222     9  11  -44
#> 135    21     259 15.5 24.44444     9  12  -21
#> 136    28     238  6.3 25.00000     9  13  -28
#> 137     9      24 10.9 21.66667     9  14   -9
#> 138    13     112 11.5 21.66667     9  15  -13
#> 139    46     237  6.9 25.55556     9  16  -46
#> 140    18     224 13.8 19.44444     9  17  -18
#> 141    13      27 10.3 24.44444     9  18  -13
#> 142    24     238 10.3 20.00000     9  19  -24
#> 143    16     201  8.0 27.77778     9  20  -16
#> 144    13     238 12.6 17.77778     9  21  -13
#> 145    23      14  9.2 21.66667     9  22  -23
#> 146    36     139 10.3 27.22222     9  23  -36
#> 147     7      49 10.3 20.55556     9  24   -7
#> 148    14      20 16.6 17.22222     9  25  -14
#> 149    30     193  6.9 21.11111     9  26  -30
#> 150    NA     145 13.2 25.00000     9  27   NA
#> 151    14     191 14.3 23.88889     9  28  -14
#> 152    18     131  8.0 24.44444     9  29  -18
#> 153    20     223 11.5 20.00000     9  30  -20

# Things transform can't do
mutate(airquality, Temp = (Temp - 32) / 1.8, OzT = Ozone / Temp)

#>     Ozone Solar.R Wind     Temp Month Day        OzT
#> 1      41     190  7.4 19.44444     5   1 2.10857143
#> 2      36     118  8.0 22.22222     5   2 1.62000000
#> 3      12     149 12.6 23.33333     5   3 0.51428571
#> 4      18     313 11.5 16.66667     5   4 1.08000000
#> 5      NA      NA 14.3 13.33333     5   5         NA
#> 6      28      NA 14.9 18.88889     5   6 1.48235294
#> 7      23     299  8.6 18.33333     5   7 1.25454545
#> 8      19      99 13.8 15.00000     5   8 1.26666667
#> 9       8      19 20.1 16.11111     5   9 0.49655172
#> 10     NA     194  8.6 20.55556     5  10         NA
#> 11      7      NA  6.9 23.33333     5  11 0.30000000
#> 12     16     256  9.7 20.55556     5  12 0.77837838
#> 13     11     290  9.2 18.88889     5  13 0.58235294
#> 14     14     274 10.9 20.00000     5  14 0.70000000
#> 15     18      65 13.2 14.44444     5  15 1.24615385
#> 16     14     334 11.5 17.77778     5  16 0.78750000
#> 17     34     307 12.0 18.88889     5  17 1.80000000
#> 18      6      78 18.4 13.88889     5  18 0.43200000
#> 19     30     322 11.5 20.00000     5  19 1.50000000
#> 20     11      44  9.7 16.66667     5  20 0.66000000
#> 21      1       8  9.7 15.00000     5  21 0.06666667
#> 22     11     320 16.6 22.77778     5  22 0.48292683
#> 23      4      25  9.7 16.11111     5  23 0.24827586
#> 24     32      92 12.0 16.11111     5  24 1.98620690
#> 25     NA      66 16.6 13.88889     5  25         NA
#> 26     NA     266 14.9 14.44444     5  26         NA
#> 27     NA      NA  8.0 13.88889     5  27         NA
#> 28     23      13 12.0 19.44444     5  28 1.18285714
#> 29     45     252 14.9 27.22222     5  29 1.65306122
#> 30    115     223  5.7 26.11111     5  30 4.40425532
#> 31     37     279  7.4 24.44444     5  31 1.51363636
#> 32     NA     286  8.6 25.55556     6   1         NA
#> 33     NA     287  9.7 23.33333     6   2         NA
#> 34     NA     242 16.1 19.44444     6   3         NA
#> 35     NA     186  9.2 28.88889     6   4         NA
#> 36     NA     220  8.6 29.44444     6   5         NA
#> 37     NA     264 14.3 26.11111     6   6         NA
#> 38     29     127  9.7 27.77778     6   7 1.04400000
#> 39     NA     273  6.9 30.55556     6   8         NA
#> 40     71     291 13.8 32.22222     6   9 2.20344828
#> 41     39     323 11.5 30.55556     6  10 1.27636364
#> 42     NA     259 10.9 33.88889     6  11         NA
#> 43     NA     250  9.2 33.33333     6  12         NA
#> 44     23     148  8.0 27.77778     6  13 0.82800000
#> 45     NA     332 13.8 26.66667     6  14         NA
#> 46     NA     322 11.5 26.11111     6  15         NA
#> 47     21     191 14.9 25.00000     6  16 0.84000000
#> 48     37     284 20.7 22.22222     6  17 1.66500000
#> 49     20      37  9.2 18.33333     6  18 1.09090909
#> 50     12     120 11.5 22.77778     6  19 0.52682927
#> 51     13     137 10.3 24.44444     6  20 0.53181818
#> 52     NA     150  6.3 25.00000     6  21         NA
#> 53     NA      59  1.7 24.44444     6  22         NA
#> 54     NA      91  4.6 24.44444     6  23         NA
#> 55     NA     250  6.3 24.44444     6  24         NA
#> 56     NA     135  8.0 23.88889     6  25         NA
#> 57     NA     127  8.0 25.55556     6  26         NA
#> 58     NA      47 10.3 22.77778     6  27         NA
#> 59     NA      98 11.5 26.66667     6  28         NA
#> 60     NA      31 14.9 25.00000     6  29         NA
#> 61     NA     138  8.0 28.33333     6  30         NA
#> 62    135     269  4.1 28.88889     7   1 4.67307692
#> 63     49     248  9.2 29.44444     7   2 1.66415094
#> 64     32     236  9.2 27.22222     7   3 1.17551020
#> 65     NA     101 10.9 28.88889     7   4         NA
#> 66     64     175  4.6 28.33333     7   5 2.25882353
#> 67     40     314 10.9 28.33333     7   6 1.41176471
#> 68     77     276  5.1 31.11111     7   7 2.47500000
#> 69     97     267  6.3 33.33333     7   8 2.91000000
#> 70     97     272  5.7 33.33333     7   9 2.91000000
#> 71     85     175  7.4 31.66667     7  10 2.68421053
#> 72     NA     139  8.6 27.77778     7  11         NA
#> 73     10     264 14.3 22.77778     7  12 0.43902439
#> 74     27     175 14.9 27.22222     7  13 0.99183673
#> 75     NA     291 14.9 32.77778     7  14         NA
#> 76      7      48 14.3 26.66667     7  15 0.26250000
#> 77     48     260  6.9 27.22222     7  16 1.76326531
#> 78     35     274 10.3 27.77778     7  17 1.26000000
#> 79     61     285  6.3 28.88889     7  18 2.11153846
#> 80     79     187  5.1 30.55556     7  19 2.58545455
#> 81     63     220 11.5 29.44444     7  20 2.13962264
#> 82     16       7  6.9 23.33333     7  21 0.68571429
#> 83     NA     258  9.7 27.22222     7  22         NA
#> 84     NA     295 11.5 27.77778     7  23         NA
#> 85     80     294  8.6 30.00000     7  24 2.66666667
#> 86    108     223  8.0 29.44444     7  25 3.66792453
#> 87     20      81  8.6 27.77778     7  26 0.72000000
#> 88     52      82 12.0 30.00000     7  27 1.73333333
#> 89     82     213  7.4 31.11111     7  28 2.63571429
#> 90     50     275  7.4 30.00000     7  29 1.66666667
#> 91     64     253  7.4 28.33333     7  30 2.25882353
#> 92     59     254  9.2 27.22222     7  31 2.16734694
#> 93     39      83  6.9 27.22222     8   1 1.43265306
#> 94      9      24 13.8 27.22222     8   2 0.33061224
#> 95     16      77  7.4 27.77778     8   3 0.57600000
#> 96     78      NA  6.9 30.00000     8   4 2.60000000
#> 97     35      NA  7.4 29.44444     8   5 1.18867925
#> 98     66      NA  4.6 30.55556     8   6 2.16000000
#> 99    122     255  4.0 31.66667     8   7 3.85263158
#> 100    89     229 10.3 32.22222     8   8 2.76206897
#> 101   110     207  8.0 32.22222     8   9 3.41379310
#> 102    NA     222  8.6 33.33333     8  10         NA
#> 103    NA     137 11.5 30.00000     8  11         NA
#> 104    44     192 11.5 30.00000     8  12 1.46666667
#> 105    28     273 11.5 27.77778     8  13 1.00800000
#> 106    65     157  9.7 26.66667     8  14 2.43750000
#> 107    NA      64 11.5 26.11111     8  15         NA
#> 108    22      71 10.3 25.00000     8  16 0.88000000
#> 109    59      51  6.3 26.11111     8  17 2.25957447
#> 110    23     115  7.4 24.44444     8  18 0.94090909
#> 111    31     244 10.9 25.55556     8  19 1.21304348
#> 112    44     190 10.3 25.55556     8  20 1.72173913
#> 113    21     259 15.5 25.00000     8  21 0.84000000
#> 114     9      36 14.3 22.22222     8  22 0.40500000
#> 115    NA     255 12.6 23.88889     8  23         NA
#> 116    45     212  9.7 26.11111     8  24 1.72340426
#> 117   168     238  3.4 27.22222     8  25 6.17142857
#> 118    73     215  8.0 30.00000     8  26 2.43333333
#> 119    NA     153  5.7 31.11111     8  27         NA
#> 120    76     203  9.7 36.11111     8  28 2.10461538
#> 121   118     225  2.3 34.44444     8  29 3.42580645
#> 122    84     237  6.3 35.55556     8  30 2.36250000
#> 123    85     188  6.3 34.44444     8  31 2.46774194
#> 124    96     167  6.9 32.77778     9   1 2.92881356
#> 125    78     197  5.1 33.33333     9   2 2.34000000
#> 126    73     183  2.8 33.88889     9   3 2.15409836
#> 127    91     189  4.6 33.88889     9   4 2.68524590
#> 128    47      95  7.4 30.55556     9   5 1.53818182
#> 129    32      92 15.5 28.88889     9   6 1.10769231
#> 130    20     252 10.9 26.66667     9   7 0.75000000
#> 131    23     220 10.3 25.55556     9   8 0.90000000
#> 132    21     230 10.9 23.88889     9   9 0.87906977
#> 133    24     259  9.7 22.77778     9  10 1.05365854
#> 134    44     236 14.9 27.22222     9  11 1.61632653
#> 135    21     259 15.5 24.44444     9  12 0.85909091
#> 136    28     238  6.3 25.00000     9  13 1.12000000
#> 137     9      24 10.9 21.66667     9  14 0.41538462
#> 138    13     112 11.5 21.66667     9  15 0.60000000
#> 139    46     237  6.9 25.55556     9  16 1.80000000
#> 140    18     224 13.8 19.44444     9  17 0.92571429
#> 141    13      27 10.3 24.44444     9  18 0.53181818
#> 142    24     238 10.3 20.00000     9  19 1.20000000
#> 143    16     201  8.0 27.77778     9  20 0.57600000
#> 144    13     238 12.6 17.77778     9  21 0.73125000
#> 145    23      14  9.2 21.66667     9  22 1.06153846
#> 146    36     139 10.3 27.22222     9  23 1.32244898
#> 147     7      49 10.3 20.55556     9  24 0.34054054
#> 148    14      20 16.6 17.22222     9  25 0.81290323
#> 149    30     193  6.9 21.11111     9  26 1.42105263
#> 150    NA     145 13.2 25.00000     9  27         NA
#> 151    14     191 14.3 23.88889     9  28 0.58604651
#> 152    18     131  8.0 24.44444     9  29 0.73636364
#> 153    20     223 11.5 20.00000     9  30 1.00000000

# mutate is rather faster than transform
system.time(transform(baseball, avg_ab = ab / g))

#>    user  system elapsed 
#>    0.01    0.00    0.01 

system.time(mutate(baseball, avg_ab = ab / g))

#>    user  system elapsed 
#>   0.000   0.000   0.001