names.quoted {plyr}	R Documentation

Compute names of quoted variables.

Description

Figure out names of quoted variables, using specified names if they exist, otherwise converting the values to character strings. This may create variable names that can only be accessed using ``.

Usage


## S3 method for class 'quoted'
names(x)

[Package plyr version 1.8.6 Index]