plyr-deprecated {plyr}	R Documentation

Deprecated Functions in Package plyr

Description

These functions are provided for compatibility with older versions of plyr only, and may be defunct as soon as the next release.

Details

liply

isplit2

[Package plyr version 1.8.7 Index]