progress_text {plyr}	R Documentation

Text progress bar.

Description

A textual progress bar

Usage

progress_text(style = 3, ...)

Arguments

style
style of text bar, see Details section of txtProgressBar

...
other arugments passed on to txtProgressBar

Details

This progress bar displays a textual progress bar that works on all platforms. It is a thin wrapper around the built-in setTxtProgressBar and can be customised in the same way.

See Also

Other progress bars: progress_none(), progress_time(), progress_tk(), progress_win()

Examples

l_ply(1:100, identity, .progress = "text")
l_ply(1:100, identity, .progress = progress_text(char = "-"))

[Package plyr version 1.8.7 Index]