progress_tk {plyr}	R Documentation

Graphical progress bar, powered by Tk.

Description

A graphical progress bar displayed in a Tk window

Usage

progress_tk(title = "plyr progress", label = "Working...", ...)

Arguments

title
window title

label
progress bar label (inside window)

...
other arguments passed on to tkProgressBar

Details

This graphical progress will appear in a separate window.

See Also

tkProgressBar for the function that powers this progress bar

Other progress bars: progress_none(), progress_text(), progress_time(), progress_win()

Examples

## Not run: 
l_ply(1:100, identity, .progress = "tk")
l_ply(1:100, identity, .progress = progress_tk(width=400))
l_ply(1:100, identity, .progress = progress_tk(label=""))

## End(Not run)

[Package plyr version 1.8.7 Index]