progress_win {plyr}	R Documentation

Graphical progress bar, powered by Windows.

Description

A graphical progress bar displayed in a separate window

Usage

progress_win(title = "plyr progress", ...)

Arguments

title
window title

...
other arguments passed on to winProgressBar

Details

This graphical progress only works on Windows.

See Also

winProgressBar for the function that powers this progress bar

Other progress bars: progress_none(), progress_text(), progress_time(), progress_tk()

Examples

if(exists("winProgressBar")) {
l_ply(1:100, identity, .progress = "win")
l_ply(1:100, identity, .progress = progress_win(title="Working..."))
}

[Package plyr version 1.8.7 Index]