split_indices {plyr}	R Documentation

Split indices.

Description

An optimised version of split for the special case of splitting row indices into groups, as used by splitter_d.

Usage

split_indices(group, n = 0L)

Arguments

n
largest integer (may not appear in index). This is hint: if the largest value of group is bigger than n, the output will silently expand.

index
integer indices

Examples

split_indices(sample(10, 100, rep = TRUE))
split_indices(sample(10, 100, rep = TRUE), 10)

[Package plyr version 1.8.7 Index]