Split indices.

An optimised version of split for the special case of splitting row indices into groups, as used by splitter_d.

split_indices(group, n = 0L)

Arguments

n
largest integer (may not appear in index). This is hint: if the largest value of group is bigger than n, the output will silently expand.

index
integer indices

Examples

split_indices(sample(10, 100, rep = TRUE))

#> [[1]]
#>  [1] 10 13 14 15 25 41 49 63 65 72 77 80 85
#> 
#> [[2]]
#>  [1]  4 11 16 20 21 27 34 35 40 55 58 92
#> 
#> [[3]]
#> [1]  5 30 59 76 87
#> 
#> [[4]]
#> [1] 37 62 64 68 70 74 84
#> 
#> [[5]]
#> [1]  6  7 19 54 67 86 93
#> 
#> [[6]]
#>  [1]  24  26  28  46  47  51  52  56  71  73  83 100
#> 
#> [[7]]
#> [1]  1 12 22 31 39 42 43 60 95
#> 
#> [[8]]
#>  [1] 32 38 44 48 53 57 66 79 81 97 98 99
#> 
#> [[9]]
#>  [1]  2  3  9 33 61 82 88 89 91 94 96
#> 
#> [[10]]
#>  [1]  8 17 18 23 29 36 45 50 69 75 78 90
#> 

split_indices(sample(10, 100, rep = TRUE), 10)

#> [[1]]
#> [1]  6 10 14 34 36 39 48 50 70
#> 
#> [[2]]
#> [1] 15 23 35 46 47 49 56 59
#> 
#> [[3]]
#>  [1]  1  2 20 38 61 65 66 68 71 90
#> 
#> [[4]]
#>  [1]  5  7 21 22 30 37 78 80 86 87
#> 
#> [[5]]
#> [1] 26 31 54 73 75 89 92
#> 
#> [[6]]
#>  [1]   3  11  12  17  29  44  52  55  57  69  74  77  84  94  97 100
#> 
#> [[7]]
#> [1] 41 45 53 72 79
#> 
#> [[8]]
#>  [1]  8 18 24 25 40 51 63 67 76 85 99
#> 
#> [[9]]
#> [1]  4 16 19 42 43 81 83 88 93
#> 
#> [[10]]
#>  [1]  9 13 27 28 32 33 58 60 62 64 82 91 95 96 98
#>