split_labels {plyr}	R Documentation

Generate labels for split data frame.

Description

Create data frame giving labels for split data frame.

Usage

split_labels(splits, drop, id = plyr::id(splits, drop = TRUE))

Arguments

splits
list of variables to split up by

drop
whether all possible combinations should be considered, or only those present in the data

[Package plyr version 1.8.9 Index]