Summarise a data frame.

Summarise works in an analogous way to mutate, except instead of adding columns to an existing data frame, it creates a new data frame. This is particularly useful in conjunction with ddply as it makes it easy to perform group-wise summaries.

summarise(.data, ...)

Arguments

.data
the data frame to be summarised

...
further arguments of the form var = value

Note

Be careful when using existing variable names; the corresponding columns will be immediately updated with the new data and this can affect subsequent operations referring to those variables.

Examples

# Let's extract the number of teams and total period of time
# covered by the baseball dataframe
summarise(baseball,
 duration = max(year) - min(year),
 nteams = length(unique(team)))

#>   duration nteams
#> 1      136    132

# Combine with ddply to do that for each separate id
ddply(baseball, "id", summarise,
 duration = max(year) - min(year),
 nteams = length(unique(team)))

#>             id duration nteams
#> 1    aaronha01       22      3
#> 2    abernte02       17      7
#> 3    adairje01       12      4
#> 4    adamsba01       20      2
#> 5    adamsbo03       13      4
#> 6    adcocjo01       16      5
#> 7    agostju01       12      5
#> 8    aguilri01       15      4
#> 9    aguirha01       15      4
#> 10   ainsmed01       14      5
#> 11   alexado01       18      8
#> 12   alexape01       19      3
#> 13   allendi01       14      5
#> 14   allenet01       12      6
#> 15   allenjo02       12      5
#> 16   almonbi01       14      7
#> 17   alomaro01       16      7
#> 18   alomasa01       14      7
#> 19   alomasa02       19      7
#> 20    aloufe01       16      7
#> 21    alouje01       16      4
#> 22    alouma01       14      6
#> 23    aloumo01       17      7
#> 24   altroni01       35      4
#> 25   alvarwi01       16      5
#> 26    amesre01       16      4
#> 27   anderbr01       14      3
#> 28   anderjo01       14      7
#> 29   anderla02       19      6
#> 30   ansonca01       26      3
#> 31   aparilu01       17      3
#> 32   appieke01       15      4
#> 33   applepe01       18      7
#> 34   applilu01       20      1
#> 35   ashburi01       14      3
#> 36   ashbyal01       16      3
#> 37   ashbyan01       13      5
#> 38   assenpa01       13      5
#> 39   astacpe01       14      8
#> 40   ausmubr01       14      3
#> 41   austiji01       20      2
#> 42   azcuejo01       12      6
#> 43   baergca01       15      6
#> 44   bagweje01       14      1
#> 45   bahnsst01       16      6
#> 46   bailebo01       16      5
#> 47   baileed01       13      5
#> 48   baineha01       21      5
#> 49    bairdo01       14      7
#> 50   bakerdu01       18      4
#> 51   bakerfl01       12      5
#> 52   bancrda01       15      4
#> 53   bandosa01       15      3
#> 54   bankser01       18      1
#> 55   bannifl01       15      6
#> 56   barbest01       14      7
#> 57   barrysh01        9      7
#> 58   bartedi01       19      5
#> 59    basske01       13      5
#> 60   baylodo01       18      6
#> 61   becklja01       19      5
#> 62   bedrost01       14      4
#> 63   belanma01       17      2
#> 64   belchti01       13      7
#> 65    bellbu01       17      4
#> 66    bellda01       11      6
#> 67    bellgu01       14      4
#> 68   bellira01       16      2
#> 69    bellja01       17      5
#> 70   benchjo01       16      1
#> 71   bendech01       22      4
#> 72   benesan01       13      4
#> 73   beniqju01       17      8
#> 74   benitar01       13      6
#> 75   bennech01       15      4
#> 76   bentola01       12      3
#> 77   bentoru01       15      2
#> 78   berenju01       14      7
#> 79   bergmda01       17      4
#> 80    bergmo01       16      5
#> 81   berrayo01       19      2
#> 82   bevacku01       14      6
#> 83   bicheda01       13      5
#> 84   bielemi01       13      5
#> 85   biggicr01       19      1
#> 86   biittla01       13      5
#> 87   blackbu02       14      5
#> 88   blairpa01       16      3
#> 89   blasido01       11      5
#> 90   bluegos01       17      1
#> 91    bluevi01       17      3
#> 92   blylebe01       22      5
#> 93   boddimi01       13      4
#> 94   boggswa01       17      3
#> 95   bondsba01       21      2
#> 96   bondsbo01       13      8
#> 97   bonilbo01       15      8
#> 98   boonebo01       18      3
#> 99   boonebr01       13      5
#> 100  boonera01       12      6
#> 101  bordepa01       17      9
#> 102  bordimi01       13      4
#> 103  bosleth01       13      7
#> 104  bottoji01       15      3
#> 105  boudrlo01       14      2
#> 106   bowala01       15      3
#> 107  bowerfr01       14      4
#> 108  boyercl02       16      3
#> 109  boyerke01       14      4
#> 110  braunst01       14      5
#> 111  bresnro01       18      5
#> 112  bressru01       18      5
#> 113  brettge01       20      1
#> 114  brettke01       14     10
#> 115  breweji01       16      3
#> 116  bridgto01       16      1
#> 117  brilene01       13      5
#> 118  brinked01       14      5
#> 119  brocklo01       18      2
#> 120  brookhu01       14      5
#> 121  broutda01       25     11
#> 122  browncl01       14      2
#> 123  brownge01       11      7
#> 124  brownha01       13      5
#> 125  brownke01       19      6
#> 126  brownmo01       13      6
#> 127  brownol02       12      6
#> 128  brownpe01       12      7
#> 129  brownto01       16     11
#> 130  brunato01       13      5
#> 131  brunege01       15     10
#> 132  bucknbi01       21      5
#> 133   buhlbo01       14      3
#> 134  buhneja01       14      2
#> 135  bunniji01       16      4
#> 136  burbada01       14      6
#> 137  burdele01       17      7
#> 138  burdoja01       19      7
#> 139  burgesm01       18      5
#> 140  burgmto01       16      5
#> 141  burkeje01       15      5
#> 142  burkejo03       16      5
#> 143  burksel01       17      5
#> 144  burnije01       13      8
#> 145  burnsge01       14      3
#> 146  burnsge02       15      5
#> 147  burrira01       14      7
#> 148  burroje01       15      6
#> 149  busbyji01       12      6
#> 150   bushdo01       15      2
#> 151   bushgu01       22      5
#> 152   bushjo01       16      7
#> 153  butlebr01       16      5
#> 154  byrneto01       14      4
#> 155  cabelen01       14      5
#> 156  caldwmi01       13      4
#> 157  callijo01       15      4
#> 158  caminke01       14      4
#> 159  campabe01       19      5
#> 160  campbbi02       14      7
#> 161  candejo01       18      8
#> 162  candito01       16      5
#> 163  cansejo01       16      7
#> 164  carbobe01       11      6
#> 165  cardejo02       17      9
#> 166  cardele01       15      5
#> 167  cardwdo01       13      5
#> 168  carewro01       18      2
#> 169  careyma01       19      2
#> 170  carlsha01       13      3
#> 171  carltst01       23      6
#> 172  carrocl02       14      6
#> 173  cartega01       18      4
#> 174  cartejo01       15      6
#> 175  cartyri01       16      7
#> 176   cashno01       16      2
#> 177  castivi02       15      6
#> 178  cavarph01       21      2
#> 179  cedence01       16      4
#> 180  cepedor01       16      6
#> 181  ceronri01       17      8
#> 182   cervbo01       11      4
#> 183    ceyro01       16      3
#> 184  chambch01       17      3
#> 185  chancfr01       16      2
#> 186  chapmbe01       16      7
#> 187  charlno01       13      6
#> 188  chaseha01       14      5
#> 189  cicoted01       15      3
#> 190  cirilje01       13      7
#> 191  clancji01       14      3
#> 192  clarkfr01       21      2
#> 193  clarkja01       17      5
#> 194  clarkwi02       14      4
#> 195  claytro01       16     11
#> 196  clemeja01       16      5
#> 197  clemero01       17      1
#> 198  clemero02       23      4
#> 199  clinety01       11      7
#> 200   cobbty01       23      2
#> 201  coffmdi01       18      5
#> 202  colavro01       13      6
#> 203  colbrgr01       12      7
#> 204  colemjo05       14      7
#> 205  colesda01       14      8
#> 206  collida02       15      8
#> 207  collied01       24      2
#> 208  colliji01       13      4
#> 209  collish01       15      2
#> 210  conceda01       18      1
#> 211   coneda01       17      5
#> 212  coninje01       17      6
#> 213  connoro01       17      5
#> 214   cookde01       14      9
#> 215  coonejo01       23      3
#> 216  coopece01       16      2
#> 217  coopewa01       17      7
#> 218  coopewi01       14      3
#> 219  corcoto01       17      5
#> 220  cordewi01       13      7
#> 221  cormirh01       16      5
#> 222  covinwe01       10      6
#> 223  cramedo01       19      4
#> 224  crandde01       17      5
#> 225  crawfsa01       18      2
#> 226  crawfwi01       13      4
#> 227  crigelo01       16      5
#> 228  cronijo01       19      3
#> 229  crosefr01       16      1
#> 230  crossla01       20      9
#> 231  crossmo01       15      5
#> 232  crowlte01       14      4
#> 233   cruzjo01       18      3
#> 234  cuccito01       15      5
#> 235  cuellmi01       18      5
#> 236  curtijo01       14      5
#> 237  cuyleki01       17      4
#> 238  dahlebi01       20      4
#> 239  dahlgba01       11      8
#> 240   dalyto01       16      5
#> 241   darkal01       14      6
#> 242  darliro01       12      3
#> 243  darwida01       20      8
#> 244  daubeja01       14      2
#> 245  daultda01       14      2
#> 246  daussho01       14      1
#> 247  davalvi01       17      6
#> 248  davisch01       18      5
#> 249  daviscu01       12      4
#> 250  daviser01       17      6
#> 251  davisge01       19      3
#> 252  davisha01       22      6
#> 253  davisma01       17      6
#> 254  davissp01       17      4
#> 255  davisst02       12      5
#> 256  davisto02       17     10
#> 257  daviswi02       19      6
#> 258  dawsoan01       20      4
#> 259  decindo01       14      3
#> 260  dejesiv01       14      7
#> 261  delahed01       15      3
#> 262  delahji01       14      8
#> 263  deleojo01       12      5
#> 264  delgaca01       14      3
#> 265  dempsri01       23      6
#> 266  dennyje01       13      7
#> 267  derripa01       14      3
#> 268  dickebi01       18      1
#> 269  dicksmu01       20      5
#> 270  dietrbi01       15      3
#> 271  dilonmi01       11      7
#> 272   doakbi01       17      3
#> 273  donovbi01       20      4
#> 274  donovdi01       15      5
#> 275  donovpa01       17      8
#> 276  dooinre01       14      3
#> 277  doolami01       13      6
#> 278  downial01       16      4
#> 279  downibr01       19      3
#> 280  doyleja01       16     10
#> 281  doylela01       13      2
#> 282  drabomo01       16      8
#> 283  driesda01       14      5
#> 284  dropowa01       12      5
#> 285  duceyro01       14      6
#> 286  duffyhu01       18      6
#> 287  duganjo01       14      5
#> 288  duncama01       12      5
#> 289  dunstsh01       17      6
#> 290  durocle01       20      4
#> 291  dwyerji01       17      7
#> 292  dykesji01       21      2
#> 293  easleda01       15      6
#> 294  easlemi01       14      6
#> 295  eckerde01       23      5
#> 296  edmonji01       14      3
#> 297  ehmkeho01       15      4
#> 298  eisenji01       16      5
#> 299  elberki01       16      6
#> 300  elliobo01       14      5
#> 301  ellisdo01       11      5
#> 302  ellswdi01       13      5
#> 303    elybo01       18      8
#> 304  embreal01       15      9
#> 305  ennisde01       13      4
#> 306  ericksc01       16      6
#> 307  etchean01       16      3
#> 308  evansda01       20      3
#> 309  evansdw01       19      2
#> 310  evereca01       13      8
#> 311  eversho01       15      5
#> 312  eversjo01       27      4
#> 313  ewingbu01       17      5
#> 314  faberre01       19      1
#> 315   facero01       16      3
#> 316  fairlro01       20      6
#> 317  farredu01       17      8
#> 318  farretu01       13      3
#> 319  fasseje01       15      9
#> 320  fellebo01       20      1
#> 321  fergujo01       13      4
#> 322  fernasi01       14      5
#> 323  fernato01       18      7
#> 324  ferreri01       18      3
#> 325  ferrewe01       14      6
#> 326  fettemi01       15      9
#> 327  fieldce01       13      5
#> 328  fingero01       17      3
#> 329  finlech01       16      4
#> 330  finlest01       18      8
#> 331  finnelo01       16      4
#> 332  fisheed02       14      6
#> 333   fiskca01       24      2
#> 334  fitzsfr01       18      2
#> 335  flagsir01       13      4
#> 336  flahejo01       13      5
#> 337  flanami01       17      2
#> 338  fletcda01       13      4
#> 339  fletcsc01       14      6
#> 340  floodcu01       15      3
#> 341  floydcl01       14      5
#> 342  foileha01       11      7
#> 343  foleyto02       12      4
#> 344   foliti01       15      6
#> 345  forceda01       15     10
#> 346   fordho01       14      5
#> 347   fordwh01       17      1
#> 348  foremfr01       18     10
#> 349  fornimi01       11      5
#> 350  forscbo01       15      2
#> 351  forscke01       16      2
#> 352  forstte01       15      5
#> 353  fostege01       17      4
#> 354  fournja01       15      5
#> 355    foxne01       18      3
#> 356   foxxji01       20      4
#> 357  francjo01       21      3
#> 358  francju01       25      8
#> 359  francti01       14      9
#> 360  frasech01       13      7
#> 361  freehbi01       15      1
#> 362  freesge02       11      6
#> 363  fregoji01       17      5
#> 364  frencla01       13      3
#> 365   freylo01       15      5
#> 366  fribebe01       14      3
#> 367  frienbo01       15      3
#> 368  friscfr01       18      2
#> 369  frymawo01       17      6
#> 370  furilca01       14      2
#> 371  gaettga01       19      6
#> 372  gagnegr01       14      3
#> 373  galanau01       15      5
#> 374  galaran01       19      7
#> 375  galehde01       15      3
#> 376  galvipu01       17      6
#> 377  gamblos01       16      7
#> 378  gantnji01       16      1
#> 379   gantro01       16      8
#> 380  ganzech01       13      4
#> 381  garbege01       19      4
#> 382  gardnla01       16      3
#> 383  garneph01       15      5
#> 384  garvene01       13      4
#> 385  garvest01       18      2
#> 386  gehrich01       18      1
#> 387  gehrilo01       16      1
#> 388  gerbewa01       15      3
#> 389  gerhajo01       18     11
#> 390  geronce01       14      3
#> 391  gibbojo01       12      4
#> 392  gibsobo01       16      1
#> 393  gibsoki01       16      4
#> 394  ginsbjo01       14      7
#> 395  girarjo01       14      4
#> 396  giustda01       15      5
#> 397  glassja01       16      9
#> 398  glavito02       20      2
#> 399  gleaski01       24      6
#> 400  gomezch02       14      7
#> 401  gonzaal01       12      6
#> 402  gonzaju03       16      4
#> 403  gonzalu01       17      5
#> 404  gonzami01       20      5
#> 405  gonzato01       11      5
#> 406  goodedw01       16      5
#> 407  goodmbi01       15      4
#> 408  goodwto01       13      6
#> 409  gordoto01       19      7
#> 410   gorege01       13      4
#> 411  gosligo01       17      3
#> 412  gossari01       22      9
#> 413   gottji01       13      4
#> 414  gowdyha01       20      2
#> 415  gracema01       15      2
#> 416  graffto01       11      6
#> 417  grantmu01       13      7
#> 418  greenle01       11      6
#> 419  greensh01       14      4
#> 420  grichbo01       16      2
#> 421  griffal01       17      4
#> 422  griffcl01       23      7
#> 423  griffke01       18      4
#> 424  griffke02       18      2
#> 425  griffto02       13      5
#> 426  grimebu01       18      7
#> 427  grimmch01       20      4
#> 428  grimsja01       17      7
#> 429  grissma02       16      6
#> 430  groatdi01       15      4
#> 431   grohhe01       15      3
#> 432  gromest01       16      2
#> 433  groombu01       13      6
#> 434  grossgr01       16      3
#> 435  grosske01       14      5
#> 436  groteje01       18      4
#> 437  grothjo01       14      5
#> 438  grovele01       16      2
#> 439  grubbjo01       15      4
#> 440  guarded01       14      3
#> 441  guerrpe01       14      2
#> 442  guilloz01       15      4
#> 443  gullibi01       15      5
#> 444  gumbeha01       15      4
#> 445   gurala01       15      3
#> 446  guthrma01       14      8
#> 447  gwynnto01       19      1
#> 448   hackst01       15      1
#> 449  haddiha01       13      5
#> 450  hadlebu01       15      6
#> 451  haineje01       19      2
#> 452  hairsje01       16      2
#> 453   halldi01       19      4
#> 454  hallmbi01       15      6
#> 455  hamilda02       13      5
#> 456  hamilea01       13      4
#> 457  hammoje01       12      6
#> 458  hamnegr01       18      3
#> 459  hanseda01       15      4
#> 460  hansero02       14      5
#> 461  hardeme01       19      1
#> 462  harnipe01       13      5
#> 463  harpebr01       16      7
#> 464  harpeto01       14      8
#> 465  harrato01       17      4
#> 466  harrebu01       15      3
#> 467  harrigr01       14      8
#> 468  harrile01       17      8
#> 469  hartnga01       19      2
#> 470  hassero01       13      6
#> 471  hasslan01       14      6
#> 472  hatchbi01       11      7
#> 473  hattogr01       14      6
#> 474  hayesch01       13      7
#> 475  hayesfr01       14      5
#> 476  haywora01       19      4
#> 477  heathcl01       14      4
#> 478  heathje01       13      4
#> 479  heathmi02       13      5
#> 480  hebneri01       17      5
#> 481  heganji01       19      5
#> 482  heganmi01       13      4
#> 483  heilmha01       18      2
#> 484   heldwo01       15      7
#> 485  helliri01       12      5
#> 486  helmsto01       13      4
#> 487  hemslro01       19      7
#> 488  hendeda01       13      5
#> 489  hendeke01       15      7
#> 490  henderi01       24      9
#> 491  hendrge01       17      6
#> 492  henrybi01       17      6
#> 493  herbera01       16      4
#> 494  hermabi01       16      4
#> 495  hernajo01       15      9
#> 496  hernake01       16      3
#> 497  hernaro01       16     10
#> 498   herrto01       12      5
#> 499  hershor01       17      4
#> 500  herzobu01       12      4
#> 501  hickmch01       11      7
#> 502  higbeki01       13      5
#> 503  higgipi01       16      3
#> 504  hillejo01       15      1
#> 505   hillgl01       12      7
#> 506   hillke01       13      7
#> 507   hillma01       13      4
#> 508  hinespa01       19     10
#> 509  hitchst01       12      4
#> 510   hoagmy01       14      4
#> 511  hodgegi01       20      3
#> 512  hoeftbi01       14      6
#> 513  hoernjo01       14      7
#> 514  hoffmtr01       14      2
#> 515  hofmaso01       13      5
#> 516  hollato01       11      8
#> 517  holmeda01       13      8
#> 518  holtzke01       14      4
#> 519  honeyri01       20      6
#> 520  hoopeha01       16      2
#> 521  hootobu01       14      3
#> 522   hoppjo01       13      6
#> 523  hornsro01       22      5
#> 524  hortowi01       17      6
#> 525  houghch01       24      4
#> 526  howarel01       13      2
#> 527  howarfr01       15      4
#> 528  howelja01       14      7
#> 529   hoytwa01       20      7
#> 530  hubbeca01       15      1
#> 531  hudliwi01       18      4
#> 532  hugheto01       13      5
#> 533  hunteca01       14      3
#> 534  hurstbr01       14      4
#> 535  incavpe01       12      6
#> 536  jacksda02       14      7
#> 537  jacksda03       14      8
#> 538  jacksgr01       17      6
#> 539  jacksla01       13      3
#> 540  jacksmi02       18      8
#> 541  jacksre01       20      5
#> 542  jackstr01       14      1
#> 543  jacobba01       12      5
#> 544  jamiech01       17      3
#> 545  jarvike01       12     10
#> 546  javiest01       17      8
#> 547  jeffegr01       13      6
#> 548  jenkife01       18      4
#> 549  jennihu01       27      6
#> 550  johnscl01       14      7
#> 551  johnsde01       16      9
#> 552  johnsja01       19      8
#> 553  johnske02       12      8
#> 554  johnsla03       13      5
#> 555  johnsra05       19      5
#> 556  johnssi01       19      4
#> 557  johnssy01       18      4
#> 558  johnswa01       20      1
#> 559   johnto01       26      6
#> 560  jonesch01       13      8
#> 561  jonesda01       17      6
#> 562  jonesdo01       18      8
#> 563  jonesfi01       19      3
#> 564  jonessa01       21      6
#> 565  jonesto02       14      8
#> 566  joneswi01       14      3
#> 567  joosted01       19      4
#> 568  jordabr01       14      4
#> 569  jorgemi01       17      6
#> 570  joynewa01       15      5
#> 571  judgejo01       19      3
#> 572  jurgebi01       16      2
#> 573  justida01       13      4
#> 574   kaatji01       24      6
#> 575  kalinal01       21      1
#> 576  keefeti01       13      5
#> 577  keelewi01       18      4
#> 578  kellejo01       17      6
#> 579   kellge01       14      5
#> 580  kellyge01       17      5
#> 581  kellyki01       15      7
#> 582  kellypa01       14      5
#> 583  kellyro01       13      8
#> 584  kennebo01       18      5
#> 585  kenneve01       11      7
#> 586   kentje01       15      6
#> 587   kernji01       12      6
#> 588  kessido01       15      3
#> 589    keyji01       14      3
#> 590  killeha01       21      3
#> 591  kingmda01       15      7
#> 592  kirkped01       15      6
#> 593  kisonbr01       14      3
#> 594  kittrma01       16      6
#> 595  kleinch01       16      3
#> 596  kleskry01       15      3
#> 597  klinero01       18      9
#> 598  klippjo01       17      8
#> 599  kluszte01       14      4
#> 600  kneppbo01       14      2
#> 601  knowlda01       15      8
#> 602  koneted01       14      6
#> 603  koosmje01       18      4
#> 604  kraneed01       17      1
#> 605  kressre01       19      5
#> 606  kreutch01       15      7
#> 607  kruegbi01       12      8
#> 608  kuennha01       14      5
#> 609  kuheljo01       17      2
#> 610  kuzavbo01       11      8
#> 611  labincl01       12      5
#> 612   lacyle01       15      4
#> 613  lajoina01       20      3
#> 614   lampde01       15      6
#> 615  landrho01       13      7
#> 616  langsma01       15      6
#> 617  laniema01       15      3
#> 618  lankfra01       14      2
#> 619  lansfca01       14      3
#> 620  lapoida01       11      9
#> 621  larkiba01       18      1
#> 622  larocda01       13      5
#> 623  larsedo01       14      8
#> 624   laryfr01       11      4
#> 625   laryly01       11      7
#> 626  lathaar01       29      7
#> 627    lauch01       11      5
#> 628  lawtoma02       11      7
#> 629    lawve01       17      1
#> 630  lazzeto01       13      4
#> 631  leachto01       20      4
#> 632  learyti01       13      7
#> 633    leebi02       13      3
#> 634    leeth01       15      3
#> 635  leibone01       12      4
#> 636  leiteal01       18      4
#> 637  lemonbo01       17      1
#> 638  lemonch01       15      2
#> 639  leonadu02       20      4
#> 640  leonaje01       13      5
#> 641  lewisda01       12      7
#> 642  leyriji01       10      6
#> 643  lindbpa01       13      5
#> 644  loaizes01       12      8
#> 645  lockmwh01       15      5
#> 646  loftoke01       16     11
#> 647  lolicmi01       16      3
#> 648  lollash01       17      4
#> 649  lombaer01       16      4
#> 650  lonboji01       14      3
#> 651   longhe01       15      5
#> 652  lopesda01       15      4
#> 653  lopezal01       19      4
#> 654  lopezja01       14      3
#> 655   lowebo01       17      4
#> 656  lowenjo01       15      3
#> 657  lowrepe01       13      4
#> 658  lucasre01       15      4
#> 659    lummi01       14      3
#> 660  luquedo01       21      4
#> 661  luzingr01       14      2
#> 662   lylesp01       15      5
#> 663   lynnfr01       16      5
#> 664  lyonste01       23      1
#> 665  mabryjo01       13      8
#> 666  macfada01       17      6
#> 667  maddoga01       14      2
#> 668  maddugr01       21      4
#> 669  maddumi01       14      9
#> 670  madlobi01       14      6
#> 671  magadda01       15      7
#> 672  mageesh01       15      3
#> 673  majesha01       16      6
#> 674  maldoca01       14      7
#> 675  mancugu01       17      5
#> 676   mannle01       15      6
#> 677  mantlmi01       17      1
#> 678  manushe01       16      6
#> 679  maranra01       23      5
#> 680  marbefi01       13      3
#> 681  maricju01       15      3
#> 682  marquru01       17      4
#> 683  marshmi01       14      9
#> 684  martibu01       17      3
#> 685  martida01       15      9
#> 686  martide01       22      5
#> 687  martied01       17      1
#> 688  martipe02       15      4
#> 689  martiti01       14      3
#> 690  martiti02       15      4
#> 691   masiph01       13      3
#> 692  mastewa02       17      3
#> 693  mathebo01       16      8
#> 694  mathech01       16      2
#> 695  matheed01       16      5
#> 696  matthga01       15      5
#> 697   maulal01       17      8
#> 698  maxvida01       13      3
#> 699  maxwech01       14      4
#> 700  maybejo01       14      4
#> 701   mayele01       12      5
#> 702    mayle01       17      4
#> 703    maymi01       14      5
#> 704  maynebr01       14      7
#> 705    mayru01       18      4
#> 706   maysca01       14      4
#> 707   mayswi01       22      3
#> 708  mazerbi01       16      1
#> 709  mazzile01       13      5
#> 710  mcauldi01       15      2
#> 711  mcbrige01       19      4
#> 712  mccarti01       21      4
#> 713  mcclubo01       18      7
#> 714  mccormi03       15      6
#> 715  mccovwi01       21      3
#> 716  mcculcl01       16      2
#> 717  mcdanli01       20      5
#> 718  mcdowsa01       14      4
#> 719  mcelrch01       12     10
#> 720  mcfared01       15      5
#> 721  mcgeewi01       17      4
#> 722  mcgrajo01       15      5
#> 723  mcgratu01       19      2
#> 724  mcgrifr01       18      6
#> 725  mcguide01       28     12
#> 726  mcgwima01       15      2
#> 727  mcinnst01       18      6
#> 728  mclemma01       18      7
#> 729  mclisca01       20      7
#> 730  mcmahdo02       17      7
#> 731  mcmanma01       14      4
#> 732  mcmilro01       15      3
#> 733  mcmulke01       15      5
#> 734  mcnaier01       13      4
#> 735  mcphebi01       17      2
#> 736  mcraeha01       19      2
#> 737  mcwilla01       12      5
#> 738  meadole01       14      3
#> 739  medwijo01       16      4
#> 740  merceor01       13      7
#> 741  merckke01       17      9
#> 742  merklfr01       19      4
#> 743   mesajo01       20      7
#> 744  micelda01       13     10
#> 745  michaca01       11      4
#> 746  miksied01       14      5
#> 747  milancl01       15      1
#> 748  millebi02       15      4
#> 749  millebo04       17     10
#> 750  milleri01       14      2
#> 751  millest01       16      6
#> 752  minchdo01       12      7
#> 753  minosmi01       31      4
#> 754  mintogr01       15      2
#> 755  mitchcl01       21      6
#> 756  mitchke01       14      8
#> 757   mizejo01       17      3
#> 758  mogrige01       16      5
#> 759  molitpa01       20      3
#> 760  mondari01       18      4
#> 761  mondera01       12      7
#> 762  moneydo01       15      2
#> 763  montawi01       16      9
#> 764  moorech02       14      2
#> 765  mooreea01       13      5
#> 766  moorege03       14      6
#> 767  moralje01       14      5
#> 768  morgajo02       21      5
#> 769  morgami01       24     12
#> 770  morrija02       17      4
#> 771  morriji01       11      5
#> 772  morrijo01       14      3
#> 773  moseswa01       16      3
#> 774   mossle01       12      4
#> 775   motama01       20      4
#> 776  moyerja01       21      7
#> 777  muellra01       16      4
#> 778  mulhote01       20     11
#> 779  mullato01       13      8
#> 780  mullige01       13      4
#> 781  mullira01       15      3
#> 782  mumphje01       14      5
#> 783  murcebo01       18      3
#> 784  murphda02       15      3
#> 785  murphda05       17      3
#> 786  murphto02       11      6
#> 787  murrada01       11      6
#> 788  murraed02       20      5
#> 789  musiast01       22      1
#> 790  mussimi01       16      2
#> 791  myattgl01       16      4
#> 792   myerbu01       16      2
#> 793  myersgr01       18      7
#> 794  myershy01       16      3
#> 795  myersmi01       12      8
#> 796  myersra01       13      6
#> 797   nashbi01       14      4
#> 798  neaglde01       12      6
#> 799   nehfar01       14      4
#> 800  nelsoca01       18      9
#> 801  nelsoje01       14      4
#> 802  nettlgr01       21      6
#> 803  nevinph01       11      7
#> 804  newhoha01       16      2
#> 805  newsobo01       24      9
#> 806  nichobi01       17      3
#> 807  nichoki01       16      3
#> 808  niekrjo01       21      7
#> 809  niekrph01       23      5
#> 810  niemabo01       11      7
#> 811  nixonot01       16      9
#> 812  norenir01       10      6
#> 813  normafr01       18      7
#> 814  northro01       15      5
#> 815  nunezed01       12      6
#> 816  nuxhajo01       22      3
#> 817  oberkke01       15      6
#> 818  obriech01       15      8
#> 819  oconnja01       23      7
#> 820   odombl01       12      5
#> 821  oeschjo01       11      4
#> 822  ofarrbo01       20      4
#> 823  offerjo01       15      7
#> 824  oglivbe01       15      3
#> 825  ojedabo01       14      5
#> 826  olerujo01       16      5
#> 827  olivaom01       11      8
#> 828  olivato01       14      1
#> 829  oliveal01       17      7
#> 830  oliveda02       14      8
#> 831  olivejo01       12      7
#> 832  olsongr01       13      9
#> 833  olsoniv01       13      3
#> 834  oneilpa01       16      2
#> 835  oneilst01       17      4
#> 836  oroscje01       24      9
#> 837  orourji01       32      8
#> 838   ortajo01       15      5
#> 839   orthal01       14      3
#> 840  osteecl01       18      6
#> 841  osterfr01       14      4
#> 842   otisam01       17      3
#> 843    ottme01       21      1
#> 844  pacioto01       17      6
#> 845  pafkoan01       16      3
#> 846  paganjo01       14      3
#> 847  palmede01       14      3
#> 848  palmeji01       19      1
#> 849  palmera01       19      3
#> 850  pappami01       16      4
#> 851  parenma01       12      7
#> 852   parkch01       13      4
#> 853  parkeda01       18      6
#> 854  parrila01       14      3
#> 855  parrila02       18      7
#> 856  pascuca02       17      6
#> 857  paskedo01       14      3
#> 858  peckiro01       17      4
#> 859  peitzhe01       21      3
#> 860   penaal01       15      6
#> 861   penaor01       17      8
#> 862   penato01       17      6
#> 863  pendlte01       14      5
#> 864  pennohe01       22      3
#> 865  perezne01       11      5
#> 866  perezto01       22      4
#> 867  perkicy01       19      3
#> 868  perraro01       12      4
#> 869  perryga01       21      8
#> 870  perryji01       16      4
#> 871  petryda01       12      4
#> 872  pfefffr01       15      5
#> 873  pfeffje01       13      4
#> 874  phillda01       21      8
#> 875  phillto02       17      7
#> 876  piazzmi01       15      5
#> 877  picinva01       17      6
#> 878  piercbi02       19      3
#> 879  piersji01       17      6
#> 880  pinielo01       20      4
#> 881  pinsova01       17      5
#> 882   pippwa01       15      3
#> 883  pizarju01       17      8
#> 884  planked01       16      3
#> 885  plesada01       17      6
#> 886  plunker01       13      4
#> 887  podrejo01       16      4
#> 888  polleho01       15      4
#> 889  polonlu01       13      6
#> 890  pooleji02       10      8
#> 891  portebo01       11      5
#> 892  porteda02       16      4
#> 893  portuma01       14      6
#> 894   postwa01       15      4
#> 895  pottene01       13      5
#> 896  powelbo01       16      3
#> 897  powelja01       15      4
#> 898  powerte01       12      8
#> 899  powervi01       11      7
#> 900  prattto02       14      4
#> 901  princto01       16      5
#> 902   puhlte01       14      2
#> 903  purcebl01       11      8
#> 904  quantpa01       13      7
#> 905  quinnja01       24      8
#> 906  quinnjo02       17      9
#> 907  quirkja01       17      8
#> 908  raffeke01       15      4
#> 909  raganpa01       14      7
#> 910  raineti01       23      6
#> 911  ramirma02       14      2
#> 912  ramospe01       15      8
#> 913  randowi01       17      6
#> 914  rasmude01       12      5
#> 915  rawlijo01       12      6
#> 916  readyra01       12      5
#> 917  reardje01       15      7
#> 918   reedje02       16      6
#> 919   reedri01       15      6
#> 920   reedro01       18      4
#> 921   reedst01       13      7
#> 922  reesepe01       18      2
#> 923  reganph01       12      4
#> 924  remlimi01       15      6
#> 925  renkost01       14      7
#> 926  reuscri01       19      4
#> 927  reussje01       21      8
#> 928  reynocr01       14      3
#> 929  rhodear01       15      5
#> 930  rhoderi01       15      4
#> 931   ricede01       16      5
#> 932   riceji01       15      1
#> 933   ricesa01       19      2
#> 934  richaha01       13      7
#> 935  richepe01       12      5
#> 936  righeda01       16      5
#> 937  ripkeca01       20      1
#> 938  rivermi01       14      3
#> 939  rixeyep01       21      2
#> 940  roberda05       12      8
#> 941  roberro01       18      4
#> 942  robinbi02       17      4
#> 943  robinbr01       22      1
#> 944  robindo01       14      4
#> 945  robined01       15      8
#> 946  robinfr02       20      5
#> 947  robinwi01       16      5
#> 948  rodriau01       16      7
#> 949  rodriiv01       16      3
#> 950  rogerke01       18      6
#> 951  rojasco01       15      4
#> 952   roofph01       16      9
#> 953   rootch01       18      2
#> 954   rosepe01       23      3
#> 955  roushed01       18      5
#> 956   rowesc01       16      3
#> 957  roystje01       15      5
#> 958   rudijo01       15      4
#> 959   ruelmu01       19      6
#> 960  ruffire01       23      3
#> 961  russebi01       17      1
#> 962  russeja01       14      6
#> 963  russeje01       13      5
#> 964   ruthba01       21      3
#> 965  ruthvdi01       13      3
#> 966   ryanji01       18      3
#> 967   ryanno01       27      4
#> 968  saberbr01       17      4
#> 969  sadecra01       17      6
#> 970  sallesl01       13      3
#> 971  samueju01       15      7
#> 972  sanchre01       14      9
#> 973  sandbry01       16      2
#> 974  sandere02       16      8
#> 975  sandesc01       18      7
#> 976  santibe01       19      9
#> 977  santoro01       14      2
#> 978  sauerha01       18      5
#> 979  schaege01       17      6
#> 980  schalra01       17      2
#> 981  schanwa01       18      5
#> 982  schatda01       14      9
#> 983  schilcu01       19      5
#> 984  schmija01       12      4
#> 985  schmijo01       15      7
#> 986  schmimi01       17      1
#> 987  schoere01       18      3
#> 988  schofdi01       18      7
#> 989  schofdi02       13      4
#> 990  schreos01       11      7
#> 991  schulfr01       14      4
#> 992  scottev01       12      5
#> 993  scottge02       13      4
#> 994  seaneru01       18      8
#> 995  seaveto01       19      4
#> 996  seguida01       14      7
#> 997  seguidi01       15      7
#> 998   seleaa01       14      6
#> 999  seminan01       14      2
#> 1000 servisc01       16      9
#> 1001 severha01       15      4
#> 1002 sewellu01       21      4
#> 1003 seymocy01       17      4
#> 1004 shaffor01       16     11
#> 1005 shantbo01       15      8
#> 1006  shawbo01       10      7
#> 1007 shawkbo01       14      2
#> 1008 sheckji01       16      6
#> 1009 sheffga01       19      7
#> 1010 sherdbi01       14      2
#> 1011 shortch02       14      2
#> 1012 shouncl01       14      5
#> 1013 siebeso01       11      6
#> 1014 sierrru01       20      9
#> 1015 sievero01       16      5
#> 1016 simmoal01       20      7
#> 1017 simmocu01       20      4
#> 1018 simmote01       20      3
#> 1019 singebi01       13      5
#> 1020 singlke01       14      3
#> 1021 sislege01       15      3
#> 1022 skowrbi01       13      5
#> 1023 slatoji01       15      3
#> 1024 slaugdo01       15      7
#> 1025 slaugen01       21      4
#> 1026 smallro02       12      4
#> 1027 smithbo02       14      3
#> 1028 smithea02       11      4
#> 1029 smithel01       15      5
#> 1030 smithge01       14      6
#> 1031 smithja03       14      2
#> 1032 smithle02       17      8
#> 1033 smithlo01       16      6
#> 1034 smithoz01       18      2
#> 1035 smithpo01       11     11
#> 1036 smithre06       16      4
#> 1037 smithsh01       16      3
#> 1038 smithza01       12      4
#> 1039 smoltjo01       19      1
#> 1040 snidedu01       17      4
#> 1041  snowjt01       14      4
#> 1042 snydefr01       15      2
#> 1043 snydepo01       18     10
#> 1044  sojolu01       13      5
#> 1045  sosasa01       18      4
#> 1046 spahnwa01       23      4
#> 1047 speaktr01       21      4
#> 1048 speiech01       18      5
#> 1049 spencji01       14      5
#> 1050 spilmha01       11      4
#> 1051 splitpa01       14      1
#> 1052 sprinru01       15      7
#> 1053 staintu01       12      7
#> 1054 stairma01       15     10
#> 1055 stalege01       14      6
#> 1056 stanlfr01       13      6
#> 1057 stanlmi01       14      1
#> 1058 stanlmi02       14      5
#> 1059 stantmi02       18      8
#> 1060 stargwi01       20      1
#> 1061 startjo01       15      6
#> 1062 staubru01       22      5
#> 1063 stengca01       13      5
#> 1064 stephve01       14      4
#> 1065 stewada01       17      5
#> 1066 stiebda01       19      2
#> 1067 stinnke01       13      8
#> 1068 stobbch01       14      5
#> 1069 stottto01       14      5
#> 1070 stoveha01       13      6
#> 1071 strawda01       16      4
#> 1072 strunam01       16      3
#> 1073 sturdto01        9      7
#> 1074 sullibi03       17      3
#> 1075 sundbji01       15      4
#> 1076 suppaje01       12      6
#> 1077 surhobj01       18      3
#> 1078 sutclri01       18      5
#> 1079 suttodo01       22      5
#> 1080 suttoez01       17      4
#> 1081 sweenma01       12      7
#> 1082 swiftbo01       13      3
#> 1083 swindgr01       16      6
#> 1084 tananfr01       20      6
#> 1085 tanneje01       17      5
#> 1086 tapanke01       12      5
#> 1087 tartada01       13      6
#> 1088 tavarju01       14      8
#> 1089 tayloto02       18      3
#> 1090 tayloza02       15      5
#> 1091 tebbebi01       16      3
#> 1092 tekulke01       15      3
#> 1093 templga01       15      3
#> 1094 tenacge01       14      4
#> 1095 tennefr02       17      2
#> 1096 terryad01       13      5
#> 1097 terryra01       11      4
#> 1098 theveto01       14      5
#> 1099 thomade01       14      7
#> 1100 thomafr03       15      7
#> 1101 thomafr04       17      3
#> 1102 thomage01       14      4
#> 1103 thomago01       13      3
#> 1104 thomato02       11      5
#> 1105 thomeji01       16      3
#> 1106 thompmi02       12      6
#> 1107 thompsa01       21      3
#> 1108 thomsbo01       14      5
#> 1109  thondi01       14      6
#> 1110 thornan01       14      3
#> 1111 tiantlu01       18      6
#> 1112 tidrodi01       12      5
#> 1113 timlimi01       16      6
#> 1114 tinkejo01       14      3
#> 1115 torgeea01       14      5
#> 1116 torrejo01       17      4
#> 1117 torremi01       17      7
#> 1118 trachst01       14      5
#> 1119 trammal01       19      1
#> 1120 traynpi01       17      1
#> 1121 trevial01       12      6
#> 1122 trillma01       16      7
#> 1123 troutdi01       18      3
#> 1124 truckvi01       17      5
#> 1125 tucketo01       12      6
#> 1126 turnete01       18      3
#> 1127  uhlege01       17      4
#> 1128 unserde01       14      5
#> 1129 valdeis01       11      7
#> 1130 valenfe01       17      6
#> 1131 valenjo03       15      5
#> 1132  valoel01       21      9
#> 1133 vanceda01       20      5
#> 1134 vandejo02       13      8
#> 1135 vanhage01       16      6
#> 1136 vaughgr01       14      5
#> 1137 veachbo01       13      4
#> 1138 velarra01       15      5
#> 1139 venturo01       15      4
#> 1140 vernomi01       21      5
#> 1141 villoro01       12     10
#> 1142 violafr01       14      5
#> 1143 vizcajo01       17      8
#> 1144 vizquom01       18      3
#> 1145 wagneha01       12      4
#> 1146 wagneho01       20      2
#> 1147 wakefti01       15      2
#> 1148 walberu01       14      3
#> 1149 walkedi02       18      5
#> 1150 walkege02       14      5
#> 1151 walkela01       16      3
#> 1152 walketo04       11      7
#> 1153 wallabo01       24      3
#> 1154 wallati01       16      3
#> 1155 wallide01       17      4
#> 1156 waltebu01       19      4
#> 1157 wanerll01       18      5
#> 1158 wanerpa01       19      4
#> 1159  wardjo01       16      4
#> 1160 warnejo01       13      7
#> 1161 warnelo01       15      2
#> 1162 washicl01       16      7
#> 1163 watsobo01       18      4
#> 1164 weathda01       16      9
#> 1165 webstmi01       12      6
#> 1166 wehmehe01       13      4
#> 1167 welchbo01       16      2
#> 1168 wellsda01       20      9
#> 1169 wertzvi01       16      6
#> 1170  westsa01       15      3
#> 1171 weyhigu01       14     10
#> 1172 wheatza01       18      2
#> 1173 whitalo01       18      1
#> 1174 whitede01       19      9
#> 1175 whitede03       16      6
#> 1176 whiteea01       16      4
#> 1177 whitefr01       17      1
#> 1178 whitema01       10      8
#> 1179 whiteri01       13     10
#> 1180 whitero01       14      1
#> 1181 whitero02       14      7
#> 1182 whitsed01       14      5
#> 1183 whitter01       15      4
#> 1184 wickmbo01       15      6
#> 1185 wightbi01       12      8
#> 1186 wilcomi01       16      5
#> 1187 wilheho01       20      9
#> 1188 wilkiri01       10      8
#> 1189 willibe02       15      1
#> 1190 willibi01       17      2
#> 1191 willicy01       18      2
#> 1192 willidi02       13      5
#> 1193 willige02       13      6
#> 1194 willima04       16      3
#> 1195 willist02       14      6
#> 1196 willite01       21      1
#> 1197 williwo02       14      4
#> 1198 willsma01       13      3
#> 1199 wilsoar01       13      6
#> 1200 wilsoji01       17      3
#> 1201 wilsowi02       18      3
#> 1202 winfida01       22      6
#> 1203 wingoiv01       18      2
#> 1204  wiseri01       18      5
#> 1205 witasja01       11      7
#> 1206  wittbo01       15      7
#> 1207 wohlfji01       14      4
#> 1208 womacto01       13      7
#> 1209  woodge01       12      8
#> 1210 woodlge01       19      6
#> 1211  woodwi01       17      3
#> 1212 worreti01       13      9
#> 1213 worthal01       16      6
#> 1214 wyattwh01       16      5
#> 1215  wynnea01       24      3
#> 1216  wynnji01       14      5
#> 1217 yastrca01       22      1
#> 1218 yeagest01       14      2
#> 1219  yorkto01       14      8
#> 1220  yosted01       18      3
#> 1221 youngcy01       21      5
#> 1222 younger01       14      7
#> 1223 youngjo02       13      5
#> 1224 yountro01       19      1
#> 1225 zachato01       18      7
#> 1226  zaungr01       12      7
#> 1227 zeileto01       15     11
#> 1228 zimmech01       19      7